% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation.R
\name{allocation}
\alias{allocation}
\alias{allocation.default}
\alias{allocation.numeric}
\alias{allocation.character}
\title{Allocation Codes}
\usage{
allocation(x, ...)

\method{allocation}{default}(x, ...)

\method{allocation}{numeric}(x, language = "english", ...)

\method{allocation}{character}(x, language = "english", ...)
}
\arguments{
\item{x}{Code.}

\item{...}{Other arguments (not used).}

\item{language}{Character string specifying language (\sQuote{english} or \sQuote{french}).}
}
\description{
Snow crab license allocation codes, used to partition landings by fisher groups.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: return allocation code table.

\item \code{numeric}: Return allocation code description.

\item \code{character}: return allocation code table.
}}

\examples{
allocation()                             # Complete allocation code table.

codes <- allocation()$code
allocation(codes[1:5])                   # Look up particular codes.
allocation(codes[1:5], language = "fr")  # Look up particular codes and return descriptions in french.
}
