% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.scsset}
\alias{as.data.frame.scsset}
\alias{as.data.frame.scsbio}
\alias{as.data.frame.probe}
\title{Convert to Data Frame}
\usage{
\method{as.data.frame}{scsset}(x, ...)

\method{as.data.frame}{scsbio}(x, ...)

\method{as.data.frame}{probe}(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Convert \strong{gulf} data objects to data frames.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Convert \code{scsset} snow crab survey biological set data object to a data frame.

\item \code{scsbio}: Convert \code{scsbio} snow crab survey biological data object to a data frame.

\item \code{probe}: Convert \code{probe} data object to a data frame.
}}

