% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\alias{catch.scsbio}
\alias{catch.scslen}
\title{Catch Data}
\usage{
catch(x, ...)

\method{catch}{scsbio}(x, by = key.scsset(), category, species, weight = FALSE, ...)

\method{catch}{scslen}(x, by = key.scslen(), species, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other parameters (not used).}

\item{by}{Character string(s) specifying which variables to group by when summararizing.}

\item{category}{Biological category string(s). See \code{\link{category}} for more details.}

\item{weight}{Logical value specifying whether to return a summary by weights rather than counts.}

\item{value}{Object containing catch data to be assigned.}
}
\description{
Extract or generate catch data summaries.
}
\section{Methods (by class)}{
\itemize{
\item \code{catch(scsbio)}: Snow crab biological data catch summaries.

\item \code{catch(scslen)}: Snow crab by-catch length data catch summaries.

}}
\examples{
s <- read.scsset(2020)  # Tow data.
b <- read.scsbio(2020)  # Biological data.

# Generate catch summaries from biological data:
catch(b) # Total crab.
catch(b, category = c("M", "F"), by = c("date"))
catch(b, category = c("COM"), by = c("date", "tow.id"))
catch(b, category = category(1:10), by = c("date", "tow.id"))
}
