% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\name{category}
\alias{category}
\alias{category.default}
\alias{category.numeric}
\alias{category.character}
\alias{is.category}
\alias{is.category.scsbio}
\alias{is.fish}
\alias{is.invertebrate}
\alias{is.skate}
\alias{is.shrimp}
\title{Biological and Species Categories}
\usage{
category(x, ...)

\method{category}{default}(x, ...)

\method{category}{numeric}(x, sex, group, ...)

\method{category}{character}(x, parse = FALSE, ...)

is.category(x, ...)

\method{is.category}{scsbio}(x, category, drop = TRUE, ...)

is.fish(species)

is.invertebrate(species)

is.skate(species)

is.shrimp(species)
}
\arguments{
\item{x}{Character string(s) specifying a biological categories or an\sQuote{scsbio} object. 
The full syntax of biological strings are presented in the \code{Details} section. =}

\item{...}{Further argument (e.g. \code{probability}) passed onto the \code{\link{is.mature.scbio}} function.}

\item{sex}{A scalar specifying the sex category of the default values to be returned if 
\code{code} is numeric. Some accepted values are \code{1}, \sQuote{m} or 
\sQuote{male} for males, or \code{2}, \sQuote{f} or \sQuote{female} for females.}

\item{parse}{Logical value specifying whether to parse a category string onto a list containing
the corresponding variable definitions, suitable for direct indexing of biological 
fields.}

\item{drop}{Logical values specifying whether to convert index to a vector from a data frame 
when only a single category is specified in \code{is.category.scsbio}.}

\item{species}{Numerical species code(s).}

\item{language}{A character string specifying the language in which category code descriptions 
are to be returned. Either english (\code{language = "english"}), 
french (\code{language = "french"} or \code{language = "français"}).}
}
\value{
Returns a vector of character strings containing the descriptions for a specified vector 
        of category strings or codes.
}
\description{
Convenience functions to quickly specify and identify biological categories and species groups. 
             These are defined by sex, size, maturity or other biological characteristics.
}
\details{
String are not case sensitive. Spaces or other separators may be included in the string for clarity, 
but have no effect on the output. String syntax is as follows:

\describe{ 
   \item{\code{Sex}}{The first character specifies the sex \sQuote{M} for males and \sQuote{F} for females.}

   \item{\code{Maturity}}{The second character(s) is/are optional and specifies the maturity. It may
                          be either \sQuote{M} for matures, \sQuote{I} for immatures, \sQuote{P} 
                          for primiparous, \sQuote{MULT} for multiparous or \sQuote{SENILE} for seniles.}

   \item{\code{Shell condition}}{Shell condition may be specified using the form \sQuote{SCX}, 
                                 where \sQuote{X} is a string of shell condition codes.}

   \item{\code{Size}}{Carapace width intervals may be specified using the forms \sQuote{GX} 
                      (greater than or equal to X) and/or \sQuote{LX} (less than X). In addition,
                      forms such as \sQuote{BTXTOY}, where \sQuote{X} and \sQuote{Y} specify 
                      lower and upper bounds, respectively. The form \sQuote{FROMXTOY} is also
                      acceptable.}
  \item{\code{Gonad colour}}{Gonad colour may be specified using the form \sQuote{GNX} where
                             \sQuote{X} is a valid colour string. Acceptable strings are \sQuote{W}
                             (White), \sQuote{B} (Beige) and \sQuote{O} (Orange).}
                             
                             
  \item{\code{Egg colour}}{Egg colour may be specified using the form \sQuote{EX} where
                           \sQuote{X} is a valid colour string. Acceptable strings are 
                           \sQuote{LO} (Light Orange), \sQuote{DO} (Dark Orange) and \sQuote{B} 
                           (Black).}

  \item{\code{Eggs remaining}}{Eggs remaining may be specified using the form \sQuote{ELXXP} 
                               or \sQuote{EGXXP}, which correspond to the less-than or 
                               greater-or-equal-to forms, respectively. The string \sQuote{X} 
                               is a two-digit percentage of eggs remaining. So crab with eggs 
                               remaining less than 50\% would be specified using the string 
                               \sQuote{EL50P}.}

  \item{\code{Missing legs}}{Crabs with missing legs may be targetted by including a \sQuote{ML} 
                             string.}
}
}
\section{Functions}{

\describe{
   \item{\code{category}}{Generic \code{category} method.}
   \item{\code{category.default}}{Default \code{category} method.}
   \item{\code{category.numeric}}{Return sets of pre-defined biological category strings.}
   \item{\code{category.character}}{Return formal and informal description or formal specification of a biological category string.}
   \item{\code{is.category}}{Generic \code{is.category} function.}
   \item{\code{is.category.scsbio}}{Returns whether an observation belongs to a specified snow crab biological category.}
   \item{\code{is.fish}}{Determine whether a species is an vertebrate fish species.}
   \item{\code{is.invertebrate}}{Determine whether a species is an invertebrate.}
   \item{\code{is.skate}}{Determine whether a fish species is a skate.}
   \item{\code{is.shrimp}}{Determine whether a species was categorized as a shrimp historically.}
}
}

\examples{
# Default biological category strings:
category()  # All.
category(sex = "female") # Female category strings.
category(1:10, sex = "male") # First 10 elements of male category strings:

# Get long form description for mature males with carapace width larger than 95mm:
category("MMG95")

# Mature males, large than 95mm with shell condition 3,4 or 5 (spaces are ignored):
category("M M G95 SC345")

# Multiparous females with orange gonads:
category("FMULTGNO")

# With multiple entries:
category(c("FM", "FMULTGNO", "MG65ML"))
   
# Parse category strings:
category("MM", parse = TRUE)         # Mature males.
category("MIG95", parse = TRUE)      # Immature males larger than 95mm.
category("FPBT35TO60", parse = TRUE) # Primiparous females between 35 and 60mm.
       
# Snow crab biological data example:
x <- read.scsbio(year = 2012)
index <- is.category(x, "TMMG95")  # Mature males greater than 95mm: 

# Get number of observations per category:
index <- is.category(x, c("T", "TM", "TF", "TMIL95SC12", "TMMG95", "TFMULT"))
apply(index, 2, sum, na.rm = TRUE)

}
