% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{check.scsbio}
\title{Data Validation}
\usage{
check(x, ...)

\method{check}{scsbio}(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Functions for validate data objects.
}
\section{Functions}{

\describe{
  \item{\code{check}}{Generic \code{check} method.}
  \item{\code{check.scsbio}}{Check \code{scsbio} object for issues.}
}
}

