% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean.scsbio}
\title{Clean or Standardize Data}
\usage{
clean(x, ...)

\method{clean}{scsbio}(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Function to remove redundant elements from a data object.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsbio}: Standardize data for \code{scsbio} object.
}}

