% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\title{Depth}
\usage{
depth(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{method}{Character string specifying which method to be used when determining water depth.}
}
\description{
Return or calculate depth
}
