% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.colour.R
\name{egg.colour}
\alias{egg.colour}
\title{Egg colour}
\usage{
egg.colour(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying egg colour code(s).}
}
\description{
Returns a description for snow crab egg colour codes. Unknown codes and NA values are returned as NA values.
}
\examples{
# Get the table of all defined egg colour codes:
egg.colour()
egg.colour(1)
egg.colour(0:2)
}
