% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.condition.R
\name{egg.condition}
\alias{egg.condition}
\title{Egg Condition}
\usage{
egg.condition(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying egg condition code(s).}
}
\description{
Returns a description for an egg condition code.

Unknown codes and NA values are returned as NA values.
}
\examples{
# Get the table of all defined egg condition codes:
egg.condition()
egg.condition(1)
egg.condition(0:2)
}
