% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg.development.R
\name{egg.development}
\alias{egg.development}
\title{Egg Development}
\usage{
egg.development(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying egg development code(s).}
}
\description{
Returns a description for snow crab egg development codes. Unknown codes and NA values are returned as NA values.
}
\examples{

egg.development()                # Table of all defined egg development codes.
egg.development(1)               # Single code.
egg.development(c(1,1,2,2,3,15)) # Vector of code values.
}
