% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.remaining.R
\name{eggs.remaining}
\alias{eggs.remaining}
\title{Eggs remaining}
\usage{
eggs.remaining(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying eggs remaining code(s).}
}
\description{
Returns a description for snow crab eggs remaining codes. Unknown codes and NA values are returned as NA values.
}
\examples{
# Get the table of all defined eggs remaining codes:
eggs.remaining(1)
eggs.remaining(0:2)
}
