% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.season.R
\name{fishing.season}
\alias{fishing.season}
\alias{fishing.season.default}
\title{Fishing Season Dates}
\usage{
fishing.season(x, ...)

\method{fishing.season}{default}(x, species = "snow crab", year, zone, ...)
}
\arguments{
\item{x}{Character search string.}

\item{...}{Other arguments (not used).}
}
\description{
Functions to retrieve fishing season date information.
}
\section{Methods (by class)}{
\itemize{
\item \code{fishing.season(default)}: Default \code{fishing.season} method. Returns the complete fishing.season data table.

}}
\examples{
fishing.season()                       # Complete fishing season table.
fishing.season(year = 2019, zone = 12) # Search for zone 12 season limits in 2019.
}
