% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing.zone.R
\name{fishing.zone}
\alias{fishing.zone}
\title{Determine Fishing Zone}
\usage{
fishing.zone(x, ...)
}
\arguments{
\item{longitude}{Numerical vector of longitudes in decimal degree format.}

\item{latitude}{Numerical vector of latitudes in decimal degree format.}

\item{region}{Character value specifying the geographic region.}

\item{species}{Numerical scalar containing a species code.}

\item{method}{Character value specifying the method to be used to determine the fishing zone.}
}
\description{
Returns a fishing zone given a set of of coordinates.
}
