% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.scobs.R
\name{fmt.scobs}
\alias{fmt.scobs}
\title{Snow crab Observer Text File Formats}
\usage{
\method{fmt}{scobs}(x, year, ...)
}
\arguments{
\item{x}{Year.}

\item{year}{Year.}
}
\description{
File format definitions to be used in reading snow crab observer data text files.
}
