% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.default}
\alias{format.default}
\alias{format<-}
\alias{format<-.default}
\title{Assign or Retrieve Data Format}
\usage{
\method{format}{default}(x, ...)

format(x, ...) <- value

\method{format}{default}(x, y, ...) <- value
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned data formats.}

\item{value}{Character string(s) specifying the data format to be assigned.}
}
\value{
Object with a \code{format} attribute attached to it.
}
\description{
Functions to assigns or retrieve data format(s) to and from an object.
}
\details{
Named argiments in the format that are not in the target object are ignored.
         The \code{format} attribute may be erased by assigning a \code{NULL} or empty 
         string \code{""} value to it.
}
\section{Functions}{

\describe{
  \item{\code{format.default}}{Default \code{format} method.}
  \item{\code{format<-}}{Generic \code{format} assignment method.}
  \item{\code{format<-.default}}{Default \code{format} assignment method. See Examples for usage.}
}
}

\examples{
x <- data.frame(year = 2010:2014, month = 6, day = 1:5, measurement = rnorm(5))
format(x, "year")  <- "YYYY"
format(x, "month") <- "MM"
format(x, "day")   <- "DD"

# Equivalently:
format(x) <- c(year = "YYYY", month = "MM", day = "DD")
format(x, c("year", "month", "day")) <- c("YYYY", "MM", "DD")

# Show attributes:
attributes(x)
format(x)

# Erase 'format' attribute:
format(x) <- NULL

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{units}}
}
