% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\alias{freq.default}
\alias{freq.scsbio}
\alias{freq.scslen}
\alias{freq.nsslen}
\title{Frequency Functions}
\usage{
freq(x, ...)

\method{freq}{default}(x, n, index, by, fill = TRUE, step, range, ...)

\method{freq}{scsbio}(x, category, by, step = 1, variable = "carapace.width", ...)

\method{freq}{scslen}(x, variable = "length", by, units = "cm", ...)

\method{freq}{nsslen}(x, variable = "length", by, units = "cm", ...)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{n}{Vector of numeric frequency values for each \code{x}.}

\item{index}{Numeric or logical indices specifying which elements of \code{x} to be analyzed as a subset.}

\item{by}{Grouping variable.}

\item{fill}{Logical value specifying whether to interpolate missing frequency bins.}

\item{step}{Numeric value specifying the width of the frequency bins.}

\item{range}{Two-element numeric vector specifying the range of values to be included in 
the resulting frequency table. Use NA or +/- Inf to specify lower or upper open bounds.}

\item{category}{Character string specifying biological categories. See \code{\link{category}} for details.}

\item{variable}{Character string specifying the variable used to generate the frequency table.}
}
\description{
Functions to generate size-frequency distributions.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{freq} method.

\item \code{scsbio}: \code{scsbio} \code{freq} method.

\item \code{scslen}: \code{scslen} \code{freq} method.

\item \code{nsslen}: \code{nsslen} \code{freq} method.
}}

\examples{
# Simple example:
x <- data.frame(values = c(11, 11, 12, 13, 15),
                n      = c(2,1,3,1,1),
                group  = c(1,1,1,2,2),
                sub    = c(1,2,2,3,3))
                
freq(x$values, x$n)  # Frequency table.
freq(x$values, x$n, by = x["group"])  # Frequency table by group.
freq(x$values, x$n, by = x[c("group", "sub")])  # Frequency table by group and subgroup.
  
# Load snow crab biological data:       
x <- read.scsbio(2020)                        

# Default frequency function examples:
freq(x$carapace.width, step = 1) # Size-frequencies using millimeter bins.
freq(x$carapace.width, step = 0.5) # Size-frequencies using half-millimeter bins.
freq(x$carapace.width, by = x["sex"], step = 0.5) # Size-frequencies by sex.
freq(x$carapace.width, by = x[c("sex", "shell.condition")], step = 0.5) # Size-frequencies by sex and shell condition.
freq(x$carapace.width, index = is.category(x, c("MM", "FM")), step = 1) # Size-frequencies for mature male and mature females.
freq(x$carapace.width, index = is.category(x, category()), step = 1)    # All default categories.

# 'scsbio' frequency function examples:
freq(x) # Size-frequencies.
freq(x, by = "sex") # Size-frequencies by sex.
freq(x, by = c("sex", "shell.condition")) # Size-frequencies by sex and shell condition.
freq(x, category = c("MM", "FM")) # Size-frequencies for mature male and mature females.
freq(x, by = c("date" "tow.id"), category = c("MM", "FM")) # Size-frequencies for mature male and mature females.

}
