% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear.R
\name{gear}
\alias{gear}
\title{Gear Codes}
\usage{
gear(x, ...)
}
\arguments{
\item{x}{Numeric gear code(s).}
}
\description{
Gear code functions.
}
\examples{
gear() # Return all codes.

gear(1)
gear(c(1, 1, 10, NA, 11, NA, 2, 2)

}
