% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glossary.R
\name{glossary}
\alias{glossary}
\alias{glossary.default}
\title{Glossary Terms for Gulf Data}
\usage{
glossary(x, ...)

\method{glossary}{default}(x, ...)
}
\arguments{
\item{x}{Search terms.}
}
\description{
Returns glossary terms and definitions used in Southern Gulf of Saint-Lawrence data sets.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Extract glossary terms.
}}

\examples{
glossary()         # Complete table of glossary terms.
glossary("moult")  # Look up 'moult' in glossary terms.
}
