% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gonad.colour.R
\name{gonad.colour}
\alias{gonad.colour}
\title{Gonad colour}
\usage{
gonad.colour(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying gonad colour code(s).}
}
\description{
Returns a description for snow crab gonad colour codes. Unknown codes and NA values are returned as NA values.
}
\examples{
# Get the table of all defined gonad colour codes:
gonad.colour(1)
gonad.colour(0:2)
}
