% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gulf.data-tables}
\alias{gulf.data-tables}
\title{Reference and Example Data Sets}
\description{
These are the reference and example data used in the \code{gulf.data} package.

Survey fishing vessel table:
\describe{
  \item{name}{Survey vessel name.}
  \item{vessel.length}{Survey vessel length.}
  \item{vessel.weight}{Vessel weight (tonnes).}
  \item{vessel.horsepower}{Survey vessel horsepower.}
  \item{vessel.hull}{Vessel hull material.}
  \item{trawl.method}{Trawl method.}
}

Survey table:
\describe{
   \item{project}{Survey project ID.}
   \item{year}{Survey year.}
   \item{id}{Survey identifier.}
   \item{vessel}{Survey vessel name.}
   \item{type}{Survey type.}
   \item{start.month}{Survey start month.}
   \item{start.day}{Survey start day.}
   \item{end.month}{Survey end month.}	
   \item{end.day}{Survey end day.}
   \item{trawl.type}{Type of trawl used.}
   \item{trawl.sensor}{Trawl acoustic sensor system.}	
   \item{minilog}{Whether a Minilog temperature-depth probe was used.}
   \item{star-oddi}{Whether Star-Oddi probe(s) were used.}
}
}
\section{\code{Minilog example.txt}}{
 
A sample Minilog text file with a header, field names, and temperature observations.

A Minilog text file with the following variables:
\describe{
  \item{date}{Date when the measurement was made.}
  \item{time}{Time when the measurement was made.}
  \item{temperature}{Temperature in degrees Celsius.}
}
}

\section{\code{species.csv}}{
 
Species Names and Codes of Atlantic fish and invertebrate species used in DFO the southern Gulf and 
Nova Scotia DFO science surveys.

A \sQuote{csv} (comma-separated) file with the following variables:
\describe{
  \item{code}{Numeric species codes used in DFO's science reasearch surveys.}
  \item{name_en}{English species names.}
  \item{name_fr}{French species names.}
  \item{name_latin}{Latin species names.}
}
}

\section{\code{species.foreign.csv}}{

Species Names and Codes of Atlantic fish and invertebrate species used by other fishery groups, such 
as the North American Fisheries Organisation and the Statistical Coordinating Committee for the Atlantic 
Coast (STACAC).

A \sQuote{csv} (comma-separated) file with the following variables:
\describe{
  \item{stacac}{Statistical Coordinating Committee for the Atlantic Coast numeric species codes.}
  \item{nafo}{North American Fisheries Organization numeric species codes.}
  \item{rv}{DFO Science research survey numeric species codes.}
  \item{name_en}{English species names.}
  \item{name_fr}{French species names.}
}
}

