% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.probe.R
\name{header.probe}
\alias{header.probe}
\alias{header.minilog}
\alias{header.star.oddi}
\alias{header.netmind}
\alias{header.esonar}
\alias{header.notus}
\title{Fetch Data Probe File Header Information}
\usage{
\method{header}{probe}(x, probe, ...)

\method{header}{minilog}(x, verbose = FALSE, ...)

\method{header}{star.oddi}(x, verbose = FALSE, ...)

\method{header}{netmind}(x, verbose = FALSE, ...)

\method{header}{esonar}(x, verbose = FALSE, ...)

\method{header}{notus}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Data file name(s).}

\item{probe}{Data probe name. See \link[gulf.data]{probe} for available options.}

\item{verbose}{Logical value specifying whether to report files as they are being read.}
}
\description{
Extract file header information from probe data files.
}
\section{Functions}{
\itemize{
\item \code{header(minilog)}: Extract Minilog file header information.

\item \code{header(star.oddi)}: Extract Star Oddi file header information.

\item \code{header(netmind)}: Extract Netmind file header information.

\item \code{header(esonar)}: Extract eSonar file header information.

\item \code{header(notus)}: Extract Notus file header information.

}}
\examples{
header.minilog()
header.minilog(1998)
}
