% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{header}
\alias{header}
\alias{header.default}
\alias{header.minilog}
\title{File Header Information}
\usage{
header(x, ...)

\method{header}{default}(x)

\method{header}{minilog}(x)
}
\arguments{
\item{x}{File name or object.}
}
\description{
Extracts header information from certain data files.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Retrieve \code{header} attribute.

\item \code{minilog}: Retrieve Minilog header information.
}}

\section{**Minilog** Header Variables}{


\describe{
  \item{Source.File}{Source fie from which the text version was generated.}
  \item{Source.Device}{Device identifier.}
  \item{Study.Description}{French species names.}
  \item{Minilog.Initialized}{Date when device was initialized.}
  \item{Study.Start.Time}{When data recording began.}
  \item{Study.Stop.Time}{When data recording stopped.}
  \item{Sample.Interval}{Sampling rate at which data were recorded.}
}
}

\examples{
file <- system.file(
header.minilog(

}
\seealso{
\code{\link[gulf.data]{minilog}}
}
