% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import<-}
\alias{import<-}
\alias{import<-.data.frame}
\alias{import<-.scsset}
\title{Import Data}
\usage{
import(x, ...) <- value

\method{import}{data.frame}(x, by, ...) <- value

\method{import}{scsset}(x, fill.na, ...) <- value
}
\arguments{
\item{x}{Target object.}

\item{...}{Other parameters (not used).}

\item{value}{Object containing catch data to be assigned.}

\item{by}{Character string(s) specifying which variables use as an index key.}
}
\description{
Import from one object into another.
}
\section{Functions}{
\itemize{
\item \code{import<-.data.frame}: Import assignment method for data frames.

\item \code{import<-.scsset}: Snow crab set import assignment method.
}}

\examples{
x <- read.scsset(year = 2020, valid = 1)
b <- read.scsbio(2020)
import(x, fill = 0) <- catch(b, category = c("M", "COM", "MF"))
print(x$M)

}
