% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.groundfish.sample.R
\name{is.groundfish.sample}
\alias{is.groundfish.sample}
\title{Groundfish Sampling?}
\usage{
is.groundfish.sample(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Returns whether a survey tow was length-sampled for groundfish.
}
\examples{
x <- read.scsset(2010:2020)
is.groundfish.sample(x)
}
