% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moulting.R
\name{is.new.shell}
\alias{is.new.shell}
\title{Carapace Moult Status}
\usage{
is.new.shell(x, ...)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical vector.
}
\description{
Determines the carapace moult or hardness status.
}
\section{Functions}{

\describe{
  \item{\code{is.new.shell}}{Generic \code{is.new.shell} method.}
  \item{\code{is.new.shell.scsbio}}{Returns whether a crab is newly-moulted based on snow crab survey data.}
  \item{\code{is.hard.shell}}{Generic \code{is.hard.shell} method.}
  \item{\code{is.hard.shell.scsbio}}{Returns whether a carapace is considered hard based on snow crab survey data.}
  \item{\code{is.hard.shell.scsobs}}{Returns whether a carapace is considered hard based on snow crab observer data.}
  \item{\code{is.soft.shell}}{Returns whether a crustacean carapace is considered soft.}
  \item{\code{is.skip.moulter}}{Generic \code{is.skip.moulter} method.}
  \item{\code{is.skip.moulter.scsbio}}{Returns whether a crustacean is considered to have skipped a moult, based on snow crab survey data.}
}
}

\examples{
# Read 2010 crab data:
x <- read.scsbio(2010)
index <- is.new.shell(x)    # Newly moulted crab.
index <- is.multiparous(x)  # Multiparous females.
index <- is.skip.moulter(x) # Skip-moulters.

}
