% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{key.default}
\alias{key<-}
\alias{key<-.default}
\alias{is.key}
\alias{is.key.data.frame}
\title{Index Key Functions}
\usage{
key(x, ...)

\method{key}{default}(x, ...)

key(x, ...) <- value

\method{key}{default}(x) <- value

is.key(x, ...)

\method{is.key}{data.frame}(x, key, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{value}{Character string(s) specifying the index key to be assigned.}
}
\description{
Functions to assign, retrieve or check an index key for a given object.
}
\section{Functions}{

\describe{
  \item{\code{key}}{Default \code{key} method.}
  \item{\code{key.default}}{Default \code{format} retrieval method. If no \code{key} attribute is 
                            defined, then a call to \code{\link[base]{format}} is made.}
  \item{\code{key<-}}{Generic \code{key} assignment method. See Examples for usage.}
  \item{\code{key<-.default}}{Default\code{key} assignment method. See Examples for usage.} 
  \item{\code{is.key}}{Default\code{is.key} method.}
  \item{\code{is.key.data.frame}}{Check validity of an index key for a \code{data.frame} object.}      
}
}

\examples{
# Build sample data:
x <- data.frame(year = 1990:2014, measurement = rpois(25))
key(x) <- "year" # Assign key.
key(x) # Retrieve key.

is.key(x, "measurement") # Generally FALSE
is.key(x, "year")
is.key(x)  # No need to specify 'year' since it is already defined.

}
\seealso{
\code{\link{metadata}}, \code{\link{description}}, \code{\link{units}}, \code{\link{format}}
}
