% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landings.R
\name{landings}
\alias{landings}
\title{Fishery Landings}
\usage{
landings(year, by)
}
\arguments{
\item{year}{Fishery year.}

\item{by}{Character vector namings variables by which to group the data.}

\item{date.start, date.end}{Start and end dates to bound the data summary.}
}
\description{
Generate a summary of fishery landings.
}
