% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.event.R
\name{locate.event}
\alias{locate.event}
\title{Find Event Time Files}
\usage{
\method{locate}{event}(x, year, project = "scs", ...)
}
\arguments{
\item{year}{Survey year.}

\item{project}{Project identifier.}
}
\description{
Locate files containing event times for various surveys.
}
