% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.logbook.R
\name{locate.logbook}
\alias{locate.logbook}
\alias{locate.ziff}
\title{Locate Fishery Logbook Data}
\usage{
\method{locate}{logbook}(x, year, species, ...)

\method{locate}{ziff}(
  x,
  year,
  path = options("gulf.path")[[1]]$groundfish$ziff,
  region,
  source = "R",
  ...
)
}
\arguments{
\item{x}{Data object.}

\item{year}{Survey year.}

\item{species}{Species to be targeted (e.g. "snow crab").}

\item{...}{Not used.}

\item{path}{Character string specifying the location of ZIFF data.}

\item{region}{DFO region (\sQuote{g}, \sQuote{m}, \sQuote{q}, \sQuote{n}).}

\item{source}{Data source (\sQuote{R}, \sQuote{ascii}).}
}
\description{
Functions to locate fishery logbook data, such as snow crab fishery data and 
             Zonal Information File Format (ZIFF) data.
}
\examples{
# Locate snow crab fishery logbook data:
locate.logbook(species = "snow crab")                        # Find all snow crab fishery logbook data files.
locate.logbook(2013:2015, species = "snow crab")             # Find all snow crab fishery logbook data files from 2013-2015.

# Locate groundfish and pelagic fishery logbook data:
locate.logbook()                                             # Find all available ziff files.
locate.ziff()                                                # Find all available ziff files.
locate.ziff(year = 2000:2008, region = "g")                  # Find Gulf region ziff files from 2000-2008.     
locate.ziff(year = 2002:2003, region = c("Gulf", "Quebec"))  # Find Gulf and Quebec regions ziff files from 2002-2003. 
locate.logbook(2000, source =  "ascii")                      # Find ASCII versions of 2000 ziff files.
}
