% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.nafo.R
\name{locate.nafo}
\alias{locate.nafo}
\title{Locate Historical NAFO Landings Data}
\usage{
\method{locate}{nafo}(x, year, source = "ascii", ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Calendar year of the landings, must be between 1960 and 1997.}

\item{source}{Character string specifying the data source for data files. Options are \sQuote{ascii} or a gulf 
package name (e.g. \sQuote{gulf.data}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}
}
\description{
Functions to locate yearly data file(s) of historical NAFO landings.
}
\examples{
locate.nafo()          # Find all NAFO files.
locate.nafo(1965)      # Find specific year.
locate.nafo(1965:1970) # Find set of years.

}
\seealso{
\code{\link{read.nafo}}
}
