% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.nss.R
\name{locate.nssset}
\alias{locate.nssset}
\alias{locate.nsscat}
\alias{locate.nsslen}
\alias{locate.nssbio}
\title{Locate Northumberland Strait Survey Data}
\usage{
\method{locate}{nssset}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{nsscat}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{nsslen}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{nssbio}(x, year, source = "gulf.data", cruise, ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{source}{Character string specifying the data source for data files. Options are \sQuote{ascii} or a gulf 
package name (e.g. \sQuote{gulf.data}).}

\item{cruise}{Character string(s) specifying a survey cruise identifier(s) (e.g. "P953").}
}
\description{
Functions to locate various data file(s) from different studies and surveys.
}
\section{Functions}{
\itemize{
\item \code{locate.nssset}: Locate Northumberland Strait survey set data files.

\item \code{locate.nsscat}: Locate Northumberland Strait survey catch data files.

\item \code{locate.nsslen}: Locate Northumberland Strait survey length data files.

\item \code{locate.nssbio}: Locate Northumberland Strait survey biological data files.
}}

\examples{
locate.nssset()
locate.nsscat()
locate.nssbio()

# Northumberland Strait length-frequency files:
locate.nsslen()
locate.nsslen(2020)

}
\seealso{
\code{\link{nss}}

\code{\link{read.nss}}
}
