% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.probe.R
\name{locate.star.oddi}
\alias{locate.star.oddi}
\alias{locate.star.oddi.default}
\alias{locate.star.oddi.scsset}
\alias{locate.scanmar}
\alias{locate.scanmar.default}
\alias{locate.esonar}
\alias{locate.esonar.default}
\alias{locate.esonar.scsset}
\title{Locate Probe Data}
\usage{
locate.star.oddi(x, ...)

\method{locate}{star.oddi.default}(
  x,
  year,
  tow.id,
  full.names = TRUE,
  probe,
  remove = c("test", "lost", "NA"),
  ...
)

\method{locate}{star.oddi.scsset}(x, ...)

locate.scanmar(x, ...)

\method{locate}{scanmar.default}(x, year, tow.id, remove = "test", ...)

locate.esonar(x, ...)

\method{locate}{esonar.default}(x, year, tow.id, remove = "test", ...)

\method{locate}{esonar.scsset}(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{tow.id}{Character string(s) specifying a snow crab survey tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{probe}{Character string specifying probe type or location. For example \sQuote{headline} specifies the Star Oddi 
probe attached to the trawl headline on snow crab surveys.}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}
}
\description{
Functions to locate probe data from different projects and surveys.
}
\section{Functions}{
\itemize{
\item \code{locate.star.oddi}: Generic method for locating Star Oddi data files.

\item \code{locate.star.oddi.default}: Default method for locating Star Oddi data files.

\item \code{locate.star.oddi.scsset}: \code{scsset} method for locating Star Oddi data files.

\item \code{locate.scanmar}: Generic method for locating Scanmar data files.

\item \code{locate.scanmar.default}: Default method for locating Scanmar acoustic data files.

\item \code{locate.esonar}: Generic method for locating eSonar data files.

\item \code{locate.esonar.default}: Default method for locating eSonar data files.

\item \code{locate.esonar.scsset}: \code{scsset} method for locating eSonar data files.
}}

\examples{
locate.esonar(2020) # Find snow crab survey eSonar data from 2020.
locate.esonar(2020, tow.id = "GP354F")

locate.star.oddi(2020, probe = "headline", source = 'ascii') 
locate.star.oddi(2020, probe = "footrope", source = 'ascii', tow.id = "GP354F")
}
