% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate.scsset}
\alias{locate.scsset}
\alias{locate.nssset}
\alias{locate.scsbio}
\alias{locate.scscat}
\alias{locate.nsscat}
\alias{locate.scslen}
\alias{locate.nsslen}
\alias{locate.nssbio}
\title{Locate Gulf Data Files}
\usage{
\method{locate}{scsset}(x, year, source = "gulf.data", ...)

\method{locate}{nssset}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{scsbio}(x, year, source = "gulf.data", remove = "bad", ...)

\method{locate}{scscat}(x, year, source = "gulf.data", remove = "bad", ...)

\method{locate}{nsscat}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{scslen}(x, year, source = "gulf.data", remove = "bad", ...)

\method{locate}{nsslen}(x, year, source = "gulf.data", cruise, ...)

\method{locate}{nssbio}(x, year, source = "gulf.data", cruise, ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{source}{Character string specifying the data source for data files. Options are \sQuote{ascii} or a gulf 
package name (e.g. \sQuote{gulf.data}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}

\item{tow.id}{Character string(s) specifying a tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{probe}{Character string specifying probe type or location. For example \sQuote{headline} specifies the Star Oddi 
probe attached to the trawl headline on snow crab surveys.}
}
\description{
Functions to locate various data file(s) from different studies and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Locate snow crab survey set data files.

\item \code{nssset}: Locate Northumberland Strait survey set data files.

\item \code{scsbio}: Locate snow crab survey biological data files.

\item \code{scscat}: Locate snow crab survey biological data files.

\item \code{nsscat}: Locate Northumberland Strait survey catch data files.

\item \code{scslen}: Locate snow crab survey biological data files.

\item \code{nsslen}: Locate Northumberland Strait survey length data files.

\item \code{nssbio}: Locate Northumberland Strait survey biological data files.
}}

\examples{
locate.scsset()          # Find all snow crab survey set data files.
locate.scsset(2020)      # Find specific year.
locate.scsset(2010:2020) # Find set of years.
 
locate.esonar(2020) # Find snow crab survey eSonar data from 2020.
locate.esonar(2020, tow.id = "GP354F")

locate.star.oddi(2020, probe = "headline", source = 'ascii') 
locate.star.oddi(2020, probe = "footrope", source = 'ascii', tow.id = "GP354F")
}
