% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.scobs.R
\name{locate.scobs}
\alias{locate.scobs}
\title{Locate Snow Crab Observer Data}
\usage{
\method{locate}{scobs}(
  x,
  year,
  path = options()$gulf.path$snow.crab$survey,
  type = "sea",
  zone,
  source = "ascii",
  ...
)
}
\arguments{
\item{x}{Data object.}

\item{year}{Fishing year.}

\item{path}{Data file path.}

\item{type}{'sea' or 'port'.}

\item{zone}{Fishing year.}

\item{source}{Data source.}
}
\description{
Functions to locate snow crab at-sea observer data file(s).
}
\examples{
locate.scobs(2010)                            # Find all at-sea sample files for the 2010 snow crab fishery.
locate.scobs(2010, type = "port")             # Find all port sample files for the 2010 snow crab fishery.
locate.scobs(2010, type = "port", zone = "E") # Find all port sample files from the 2010 snow crab fishery in zone 12E.

}
