% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.scs.R
\name{locate.scsset}
\alias{locate.scsset}
\alias{locate.scscat}
\alias{locate.scsbio}
\alias{locate.scslen}
\title{Locate Snow Crab Survey Data}
\usage{
\method{locate}{scsset}(x, year, source = "gulf.data", ...)

\method{locate}{scscat}(x, year, source = "gulf.data", remove = "bad", ...)

\method{locate}{scsbio}(x, year, source = "gulf.data", remove = "bad", ...)

\method{locate}{scslen}(x, year, source = "gulf.data", remove = "bad", ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Survey year.}

\item{source}{Character string specifying the data source for data files. Options are \sQuote{ascii} or a gulf 
package name (e.g. \sQuote{gulf.data}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}

\item{tow.id}{Character string(s) specifying a tow identifier(s) (e.g. \sQuote{GP354F}).}
}
\description{
Functions to locate various data file(s) from the Snow Crab Survey.
}
\section{Functions}{
\itemize{
\item \code{locate(scsset)}: Locate snow crab survey set data files.

\item \code{locate(scscat)}: Locate snow crab survey biological data files.

\item \code{locate(scsbio)}: Locate snow crab survey biological data files.

\item \code{locate(scslen)}: Locate snow crab survey by-catch length data files.

}}
\examples{
locate.scsset()          # Find all snow crab survey set data files.
locate.scsset(2020)      # Find specific year.
locate.scsset(2010:2020) # Find set of years.

}
\seealso{
\code{\link{scs}}

\code{\link{read.scs}}
}
