% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{maturity}
\alias{maturity}
\alias{is.mature}
\alias{is.mature.scsbio}
\alias{is.primiparous}
\alias{is.primiparous.scsbio}
\alias{is.multiparous}
\alias{is.multiparous.scsbio}
\alias{is.senile}
\alias{is.senile.scsbio}
\alias{maturity.default}
\alias{maturity.numeric}
\title{Sexual Maturity Functions}
\usage{
is.mature(x, ...)

\method{is.mature}{scsbio}(x, probability = FALSE, ...)

is.primiparous(x, ...)

\method{is.primiparous}{scsbio}(x, ...)

is.multiparous(x, ...)

\method{is.multiparous}{scsbio}(x, ...)

is.senile(x, ...)

\method{is.senile}{scsbio}(x, ...)

maturity(x, ...)

\method{maturity}{default}(x, ...)

\method{maturity}{numeric}(x, year, species, ...)
}
\arguments{
\item{x}{Object or numeric maturity code.}

\item{...}{Further arguments passed onto internal functions.}

\item{probability}{Logical value specifying whether values are to be returned as 
probabilities of being mature. For individuals with missing morphometric 
data or observational sexual characters, a GAM binary regression is 
performed on the observations by year and the predicted probabilties are 
returned.}

\item{year}{Survey year.}

\item{species}{Species name or code.}
}
\value{
Logical vector, probability (when \code{probability = TRUE}), or character string.
}
\description{
Determines whether an organism is sexually mature, based on qualitative or 
             morphometric observational data.
}
\section{Functions}{

\describe{
  \item{\code{is.mature}}{Generic \code{is.mature} function.}
  \item{\code{is.mature.scsbio}}{Snow crab biological data (\code{scsbio}) maturity method.}
  \item{\code{is.primiparous}}{Generic \code{is.primiparous} function.}
  \item{\code{is.primiparous.scsbio}}{Returns whether a female is primiparous, i.e. a first-time spawner.}
  \item{\code{is.multiparous}}{Generic \code{is.multiparous} function.}
  \item{\code{is.multiparous.scsbio}}{Returns whether a female is multiparous, i.e. has previously spawned.}
  \item{\code{is.senile}}{Generic \code{is.senile} function.}
  \item{\code{is.senile.scsbio}}{Returns whether an animal is senile, i.e. has decreased spawning capability because of old age.}
  \item{\code{maturity}}{Generic \code{weight} method.}
  \item{\code{maturity.default}}{Return sexual maturity status.}
  \item{\code{maturity.numeric}}{Convert numeric codes to descriptive maturity character strings.}
}
}

\examples{
x <- read.scsbio(2010)
index <- is.mature(x)
plot(x$carapace.width[index], x$chela.height[index], xlim = c(0, 140), ylim = c(0, 40), 
     xlab = "Carapace width (mm)", ylab = "Chela height (mm)", type = "n")
points(x$carapace.width[index], x$chela.height[index], pch = 21, bg = "grey", cex = 0.4)
points(x$carapace.width[!index], x$chela.height[!index], pch = 21, bg = "red", cex = 0.4)

}
