% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum.legal.size.R
\name{minimum.legal.size}
\alias{minimum.legal.size}
\alias{minimum.legal.size.default}
\title{Minimum Legal Size}
\usage{
minimum.legal.size(x, ...)

\method{minimum.legal.size}{default}(x, species, year, fishing.zone, sub.fishing.zone, ...)
}
\description{
Return the minimum legal sizes.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default minimum legal size function.
}}

