% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.legs.R
\name{missing.legs}
\alias{missing.legs}
\alias{missing.legs.default}
\alias{missing.legs.numeric}
\alias{missing.legs.character}
\alias{is.missing.legs}
\alias{is.missing.legs.character}
\alias{is.missing.legs.data.frame}
\alias{bin2str.missing.legs}
\title{Missing Legs}
\usage{
missing.legs(x, ...)

\method{missing.legs}{default}(x)

\method{missing.legs}{numeric}(x)

\method{missing.legs}{character}(x)

is.missing.legs(x, ...)

\method{is.missing.legs}{character}(x, side, include.regenerated = FALSE, ...)

\method{is.missing.legs}{data.frame}(x, side, include.regenerated = FALSE, ...)

bin2str.missing.legs(x)
}
\arguments{
\item{x}{Numerical code(s), character string(s) or \code{list} objects.}
}
\value{
Character string(s) of logical values.
}
\description{
Functions for formatting and checking whether a crustacean has missing legs.
}
\section{Standard missing legs format}{


The standard missing leg format is a 10-character string with the proper missing legs codes.
The first five characters correspond to the left-hand side pereiopds, numbered starting with the chela. 
The last 5 characters correspond to those of the right-hand side. 

For example \sQuote{**2***1***} indicates that the 3rd leg on the left-hand side is regenerated,
while the 2nd leg on the right-hand side is missing.
}

\section{Functions}{

\describe{
  \item{\code{missing.legs}}{Generic \code{missing.legs} method.}
  \item{\code{missing.legs.default}}{Default \code{missing.legs} method.}
  
  \item{\code{missing.legs.numeric}}{Returns a missing legs description for a specified missing legs code. 
                                     If no \code{code} is given, then a named character vector of all missing 
                                     legs names for a given survey is returned. Unknown codes and NA values are 
                                     returned as NA values.}
  
  \item{\code{missing.legs.character}}{Function to convert non-standard character string input to the standard missing 
                                       legs format. The function identifies certain character string fragments as relevant
                                       for missing legs codes. For example, \sQuote{"R2M L3R"} specifies that the 2nd
                                       leg on the right-hand side is missing while the 3rd leg on the left-hand
                                       side is regenerated. This string translates to \sQuote{**2***1***} in the
                                       standard format. It is also possible to use a more compact coding available such as
                                       \sQuote{"R23 L12R3R"}, which specifies that the 2nd and 3rd legs on the
                                       right-hand side is missing, and the first leg on the left-hand side is
                                       missing, while the 2nd and 3rd ones are regenerated. This string translates
                                       to \sQuote{122***11**} in the standard format.}
                                       
  \item{\code{is.missing.legs}}{Generic \code{is.missing.legs} function.}
  
  \item{\code{is.missing.legs.character}}{Returns whether a crustacean has missing legs.}
  
  \item{\code{is.missing.legs.list}}{Returns whether a crustacean has missing legs for an given object containing missing
                                     leg data observations.}
}
}

\examples{
# Return all missing legs code descriptions:
missing.legs()
   
# Missing legs descriptions for specified codes:
missing.legs(c(1, 1, 2, 3)) 

# A sample of test cases:
str <- c("11 L234 R34R", "15 R2M L3R L4M L5", "  ", "", "56", "r3",
         "15 L2,3,5,R2,4", "That's R2 right", "R234  L2,3r", "l3r", " 34 what",
         "L1,3,2R", "This a comment", "R234r5r, l12r3, 49")

# Convert to standard format:
missing.legs(str)

# Read 2010 crab data:
x <- read.scsbio(year = 2010)
   
# Remove specimens with missing legs:
x <- x[is.missing.legs(x), ]

}
