% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moulting.R
\name{moulting}
\alias{moulting}
\alias{is.new.shell}
\alias{molting}
\alias{is.new.shell.scsbio}
\alias{is.hard.shell}
\alias{is.hard.shell.scsbio}
\alias{is.hard.shell.scobs}
\alias{is.soft.shell}
\alias{is.skip.moulter}
\alias{is.skip.moulter.scsbio}
\title{Carapace Moult Status}
\usage{
is.new.shell(x, ...)

\method{is.new.shell}{scsbio}(x)

is.hard.shell(x, ...)

\method{is.hard.shell}{scsbio}(x, durometer, shell.condition, zone, ...)

\method{is.hard.shell}{scobs}(x, durometer, shell.condition, zone, ...)

is.soft.shell(x, ...)

is.skip.moulter(x, ...)

\method{is.skip.moulter}{scsbio}(x, ...)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical vector.
}
\description{
Determines the carapace moult or hardness status.
}
\section{Functions}{

\describe{
  \item{\code{is.new.shell}}{Generic \code{is.new.shell} method.}
  \item{\code{is.new.shell.scsbio}}{Returns whether a crab is newly-moulted based on snow crab survey data.}
  \item{\code{is.hard.shell}}{Generic \code{is.hard.shell} method.}
  \item{\code{is.hard.shell.scsbio}}{Returns whether a carapace is considered hard based on snow crab survey data.}
  \item{\code{is.hard.shell.scobs}}{Returns whether a carapace is considered hard based on snow crab observer data.}
  \item{\code{is.soft.shell}}{Returns whether a crustacean carapace is considered soft.}
  \item{\code{is.skip.moulter}}{Generic \code{is.skip.moulter} method.}
  \item{\code{is.skip.moulter.scsbio}}{Returns whether a crustacean is considered to have skipped a moult, based on snow crab survey data.}
}
}

\examples{
# Read 2020 crab data:
x <- read.scsbio(2020)
index <- is.new.shell(x)    # Newly moulted crab.
index <- is.soft.shell(x)   # Soft-shelled crab.
index <- is.skip.moulter(x) # Skip-moulters.

}
