% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmind.R
\name{netmind}
\alias{netmind}
\alias{netmind.default}
\title{\strong{Netmind} Data}
\usage{
netmind(x, ...)

\method{netmind}{default}(x, ...)
}
\arguments{
\item{x}{An \code{netmind} object, data file, survey year or keyword search term.}

\item{header}{\strong{Netmind} file header information to be assigned as metadata.}
}
\description{
Create \strong{Netmind} trawl acoustic monitoring probe data objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{netmind(default)}: Create an \code{netmind} object.

}}
\examples{
# Netmind files for the 2020 snow crab survey:
locate.netmind(year = 1999)

# Locate files with a specific tow ID from snow crab survey 2018-2020:
locate.netmind(1999, tow.id = 223)

# Working example:
locate.netmind(1999, tow.id = 223)

}
\seealso{
\link{scanmar}, \link{esonar}, \link{notus}
}
