% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notus.R
\name{notus}
\alias{notus}
\alias{notus.default}
\title{\strong{Notus} Data}
\usage{
notus(x, ...)

\method{notus}{default}(x, ...)
}
\arguments{
\item{x}{An \code{notus} object, data file, survey year or keyword search term.}

\item{header}{\strong{Notus} file header information to be assigned as metadata.}
}
\description{
Create \strong{Notus} trawl acoustic monitoring probe data objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{notus(default)}: Create an \code{notus} object.

}}
\examples{
# Notus files for the 2020 snow crab survey:
locate.notus(year = 2020, project == "nss")

}
\seealso{
\link{scanmar}, \link{netmind}, \link{esonar}
}
