% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nss.R
\name{nssset}
\alias{nssset}
\alias{nssset.default}
\alias{nsscat.default}
\alias{nssbio.default}
\alias{nsslen.default}
\title{Northumberland Strait Survey Data Classes}
\usage{
nssset(x, ...)

\method{nssset}{default}(x, ...)

\method{nsscat}{default}(x, ...)

\method{nssbio}{default}(x, ...)

\method{nsslen}{default}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object containing data fields to be mapped onto a corresponding snow crab survey data object.}

\item{...}{Other parameters (not used).}
}
\description{
Class containers for snow crab survey data.
}
\section{Functions}{
\itemize{
\item \code{nssset}: Generic \code{nssset} method.

\item \code{nssset.default}: Create an \code{nssset} object.

\item \code{nsscat.default}: Create an \code{nsscat} object.

\item \code{nssbio.default}: Create an \code{nssbio} object.

\item \code{nsslen.default}: Create an \code{nsslen} object.
}}

\examples{
# Create 'nsscat' object from a data frame:
x <- nsscat(data.frame(date = "2020-07-19", cruise = "P555", set.number = 1, species = 10, number.caught = 144))
key(x)

# Create 'nssbio' object from a data frame:
x <- nssbio(data.frame(date = "2020-07-19", cruise = "P555", set.number = 1, species = 10, specimen = 1, length = 10))
key(x)

# Create 'nsslen' object from a data frame:
x <- nsslen(data.frame(date = "2020-07-19", cruise = "P555", set.number = 1, species = 10, size.class = 1, sex = 1, length = 10))
key(x)

}
\seealso{
\code{\link{read.nssset}},  \code{\link{read.nsscat}}
}
