% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nssbio.R
\name{nssbio}
\alias{nssbio}
\alias{nssbio.default}
\title{Northumberland Strait Survey Biological Data}
\usage{
nssbio(x, ...)

\method{nssbio}{default}(x, ...)
}
\arguments{
\item{x}{A \sQuote{data.frame} object. When reading data, \code{x} may be a numeric vector specifying the survey years.}

\item{year}{Survey year(s) to be loaded.}
}
\description{
The \code{nssbio} class is a container for Northumberland Strait survey biological data.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{nssbio} object.
}}

\examples{
# Create 'nssbio' object from a data frame:
x <- nssbio(data.frame(date = "2020-07-19", cruise = "P555", set.number = 1, species = 10, specimen = 1, length = 10))
key(x)

# Read data:    
x <- read.nssbio()                 # Read all available data.
x <- read.nssbio(year = 2019)      # Read single year.
x <- read.nssbio(year = 2010:2015, species = 10) # Load cod data for range of years.
x <- read.nssbio(species = "Cod") # Load cod data using species name search for all years.

}
\seealso{
\code{\link{read.nssset}}, \code{\link{read.nsscat}}, \code{\link{read.nsslen}}
}
