% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsscat.R
\name{nsscat}
\alias{nsscat}
\alias{nsscat.default}
\title{Northumberland Strait Survey Set/Tow Data}
\usage{
nsscat(x, ...)

\method{nsscat}{default}(x, ...)
}
\arguments{
\item{x}{A \sQuote{data.frame} object. When reading data, \code{x} may be a numeric vector specifying the survey years.}

\item{year}{Survey year(s) to be loaded.}
}
\description{
The \code{nsscat} class is a container for Northumberland Strait survey set/tow data, 
             i.e. detailed information about individual tows performed during the annual Northumberland Strait
             survey.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{nsscat} object.
}}

\seealso{
\code{\link{read.nsscat}}
}
