% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsslen.R
\name{nsslen}
\alias{nsslen}
\alias{nsslen.default}
\title{Northumberland Strait Survey Set/Tow Data}
\usage{
nsslen(x, ...)

\method{nsslen}{default}(x, ...)
}
\arguments{
\item{x}{A \sQuote{data.frame} object. When reading data, \code{x} may be a numeric vector specifying the survey years.}

\item{year}{Survey year(s) to be loaded.}
}
\description{
The \code{nsslen} class is a container for Northumberland Strait survey length data.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{nsslen} object.
}}

\examples{
# Create 'nsslen' object from a data frame:
x <- nsslen(data.frame(date = "2020-07-19", cruise = "P555", set.number = 1, species = 10, number.caught = 144))
key(x)

# Read data:    
x <- read.nsslen()                 # Read all available data.
x <- read.nsslen(year = 2019)      # Read single year.
x <- read.nsslen(year = 2010:2015, species = 10) # Load cod data for range of years.
x <- read.nsslen(species = "Cod") # Load cod data using species name search for all years.

}
\seealso{
\code{\link{read.nssset}}, \code{\link{read.nsscat}}
}
