% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.minilog.R
\name{plot.minilog}
\alias{plot.minilog}
\title{Plot Minilog Data}
\usage{
\method{plot}{minilog}(x, col, lwd = 1, ...)
}
\arguments{
\item{x}{Minilog object.}

\item{...}{Line style arguments (see \code{\link[graphics]{lines}}).}
}
\description{
Plots the data recorded by a Minilog probe, generally the temperature and water depth.
}
\examples{
file <- system.file("extdata", "Minilog example.txt", package = "gulf.data")
x <- read.minilog(file)
plot(x)

}
