% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R
\name{probe}
\alias{probe}
\alias{probe.default}
\alias{probe.character}
\alias{probe.data.frame}
\title{Project Identifiers}
\usage{
probe(x, ...)

\method{probe}{default}(x, ...)

\method{probe}{character}(x, ...)

\method{probe}{data.frame}(x, header, ...)
}
\arguments{
\item{x}{Character search string.}

\item{...}{Other arguments (not used).}

\item{header}{File header information.}
}
\description{
Functions to retrieve survey probe information or create a probe data object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Load data probe description table.

\item \code{character}: Find a data probe by name.

\item \code{data.frame}: Create a \code{probe} class object.
}}

\examples{
probe()            # Load complete data probe table.
probe("min")       # Search for probe name.
}
