% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\alias{project.gulf.set}
\title{Project Identifier for Gulf Data}
\usage{
\method{project}{gulf.set}(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{...}{Other arguments.}
}
\description{
Functions to retrieve project identifiers for \sQuote{gulf.set}, \sQuote{gulf.cat},
             \sQuote{gulf.bio} and \sQuote{gulf.len} objects.
}
