% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.alsi.R
\name{read.alsi}
\alias{read.alsi}
\title{Read Atlantic Lobster Settlement Index Data}
\usage{
read.alsi(year, site, species, type, ...)
}
\arguments{
\item{year}{Project year.}

\item{site}{Character string specifying site name.}

\item{species}{Species code or name.}

\item{type}{Type of data to be read.}

\item{x}{Survey year or file name.}
}
\description{
Functions to read probe data, such as depth/temperature or acoustic trawl monitoring data.
}
\section{Functions}{
\itemize{
\item \code{read.alsi()}: Read data from the Altantic Lobster Stock-recruitment Index. Larvae collectors are used to mainly provide recruitment indices for lobster and rock crab.

}}
