% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.azmp.R
\name{read.azmp}
\alias{read.azmp}
\title{Read AZMP data}
\usage{
read.azmp(x, file, station, year, month, ...)
}
\arguments{
\item{x}{Year or file name.}

\item{file}{File name(s)}

\item{station}{AZMP station identifier.}

\item{year}{Year.}

\item{month}{Month>}
}
\value{
Returns a data.frame with date, time, latitude and longitude fields.
}
\description{
This function reads Atlantic Zonal Monitoring Program (AZMP) oceanographic data.
}
\seealso{
\code{\link{probe}}
}
