% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.collector.R
\name{read.collector}
\alias{read.collector}
\title{Lobster Larvae Collector Data}
\usage{
read.collector(year, site, species, type, ...)
}
\description{
Larvae collectors are used to provide recruitment indices for lobster and rock crab. 
This function provides access to these data.
}
\examples{
x <- read.collector(year = 2010, species = 2550)

}
