% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.colorimeter.R
\name{read.colorimeter}
\alias{read.colorimeter}
\alias{locate.colorimeter}
\title{Locate and Read Colorimetry Data}
\usage{
read.colorimeter(year = NULL, ...)

\method{locate}{colorimeter}(x, year, source = "gulf.data", ...)
}
\arguments{
\item{year}{Survey year.}

\item{x}{Data object.}

\item{tow.id}{Character string(s) specifying a tow identifier(s) (e.g. \sQuote{GP354F}).}
}
\description{
Functions to find and access colour meter data from the snow crab survey.
}
\section{Methods (by generic)}{
\itemize{
\item \code{locate}: Locate snow crab colorimetry files.
}}

\examples{
locate.colorimeter()          # Find all snow crab colorimeter data files.
locate.colorimeter(2020)      # Find specific year.
locate.colorimeter(2017:2019) # Find set of years.

}
\seealso{
\code{\link{read.colorimeter}}
}
