% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ctd.R
\name{read.ctd}
\alias{read.ctd}
\title{Read CTD data}
\usage{
read.ctd(
  year,
  survey = c("rvs", "scs", "nss"),
  table = "ctd",
  dsn = "SGDOP",
  username = "SGDO_lecture",
  password,
  echo = TRUE
)
}
\arguments{
\item{year}{Survey or study year.}

\item{survey}{Character string specifying which survey to target. Available codes are the Septmeber survey (\code{rvs}), the snow crab survey (\code{scs}), 
and the Northumberland Strait survey (\code{nss}). The default returns data from all available surveys.}

\item{table}{Name of data table (\code{ctd}, \code{bottle} or \code{metadata}).}

\item{dsn}{Oracle data source name (= \code{SGDOP}).}

\item{username}{Oracle user name (= \code{SGDO_lecture}).}

\item{password}{Oracle password.}

\item{echo}{Logical value specifying whether to return messages during data retrieval.}
}
\description{
Read Conductivity-Temperature-Depth (CTD) data directly from the Institut Maurice-Lamontagne (IML) SGDOP data base. 
In particular, the database contains the oceanographic data collected during science surveys in the southern Gulf
of St. Lawrence. Oracle client needs to be set up in ODBC sources.
}
