% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.event.R
\name{read.event}
\alias{read.event}
\title{Find Event Time Files}
\usage{
read.event(x, ...)
}
\arguments{
\item{year}{Survey year.}

\item{project}{Project identifier.}
}
\description{
Locate files containing event times for various surveys.
}
