% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gps.R
\name{read.gps}
\alias{read.gps}
\title{Read GPS Data}
\usage{
read.gps(x)
}
\arguments{
\item{x}{File name(s).}
}
\value{
Returns a data.frame with date, time, latitude and longitude fields.
}
\description{
This function reads a GPS track file.
}
\seealso{
\code{\link{probe}}
}
