% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gulf.cat.R
\name{read.gulf.cat}
\alias{read.gulf.cat}
\title{Read Survey Catch Data}
\usage{
read.gulf.cat(
  year,
  species,
  survey = "rv",
  source = "oracle",
  dsn = options("gulf.oracle")[[1]]$rvs$dsn,
  uid = options("gulf.oracle")[[1]]$rvs$uid,
  ...
)
}
\arguments{
\item{year}{Survey year.}

\item{species}{Species name or code.}

\item{survey}{Survey name.}

\item{dsn}{Data source name.}

\item{uid}{User identification.}

\item{password}{Oracle database password.}
}
\description{
Read trawl catch data from various groundfish, crustacean and pelagic surveys.
}
