% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gulf.set.R
\name{read.gulf.set}
\alias{read.gulf.set}
\title{Read Survey Data}
\usage{
read.gulf.set(
  year,
  survey = "rv",
  source = "oracle",
  dsn = options("gulf.oracle")[[1]]$rvs$dsn,
  uid = options("gulf.oracle")[[1]]$rvs$uid,
  ...
)
}
\arguments{
\item{year}{Survey year.}

\item{survey}{Survey name.}

\item{source}{Data source name.}
}
\description{
Read trawl set data from various groundfish and lobster surveys.
}
