% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.logbook.R
\name{read.logbook}
\alias{read.logbook}
\alias{read.ziff}
\alias{read.sc.logbook}
\title{Load Fishery Logbook Data}
\usage{
read.logbook(x, species = "snow crab", ...)

read.ziff(x, echo = TRUE, ...)

read.sc.logbook(
  x,
  year,
  file,
  path = options("gulf.path")[[1]]$snow.crab$logbook,
  ...
)
}
\arguments{
\item{x}{Fishery year or file name.}

\item{echo}{Logical value specifying whether to report files to the R console as they're being read.}

\item{year}{Fishery year.}

\item{file}{File name to be loaded. File format is assumed to be comma-separated (i.e. .csv).}

\item{path}{File path containing comma-separated fishery logbook files (i.e. .csv).}
}
\description{
Function to load logbook and slip landings data from the southern Gulf of Saint Lawrence fisheries.
}
