% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.minilog.R
\name{read.minilog}
\alias{read.minilog}
\alias{write.minilog}
\alias{path.minilog}
\alias{file.minilog}
\title{Read and Write Minilog Data}
\usage{
read.minilog(x, project, offset = 0, file, ...)

write.minilog(x, file = NULL)

path.minilog(year, project, survey, ...)

file.minilog(
  x,
  year,
  set.number,
  tow.id,
  set.card,
  path = TRUE,
  survey = "sc",
  project,
  ...
)
}
\arguments{
\item{x}{Minilog data file}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{survey, project}{Character string specifying the survey or project name. 
The survey may refer to the September multi-species survey (\code{survey = "rv"},
\code{survey = "sep"} or \code{survey = "September"}), the  Northumberland Strait 
survey (\code{= "ns"}), the mobile Sentinel survey (\code{= "sen"} or \code{= "sentinel"}),
or the snow crab survey (\code{= "sc"} or \code{= "snow crab"}).}

\item{set.number, tow.id}{Numeric value or strings specifying the survey set number or tow ID.}

\item{set.card}{Data frame containing the survey tows which specify the Minilog data to be loaded.}

\item{path}{Logical value specifying whether to include the path in the Minilog files.}
}
\description{
Minilog probes are used to record depth and temperature data. 
These are functions to read and write \code{minilog} objects.
}
\section{Functions}{
\itemize{
\item \code{write.minilog}: Write a \code{minilog} object to a file.

\item \code{path.minilog}: Find a minilog data path.

\item \code{file.minilog}: Retrieve file name of a Minilog dataset.
}}

\examples{
file <- system.file("extdata", "Minilog example.txt", package = "gulf.data")
x <- read.minilog(file)
header(x)
plot(x)

# Use a set card extract for the first ten sets of 2012:
x <- read.gulf(year = 2012, survey = "sc")
minilog.file(x[1:10, ])

# Use a tow ID to extract file names for the snow crab survey 2006-2012:
minilog.file("GP001", year = 2006:2012)

x <- read.minilog(tow.id = "GP001F", year = 2010)
x <- read.minilog(survey = "rv", year = 2010)

}
\seealso{
\code{\link[gulf.data]{minilog}}
}
