% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nafo.R
\name{read.nafo}
\alias{read.nafo}
\title{Read NAFO Landings Data}
\usage{
read.nafo(x, file, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{year}{Survey year(s).}
}
\value{
A data frame is returned.
}
\description{
Read a fixed-width formatted NAFO landings data file.
}
\examples{

   # Read the 1960-1965 NAFO landings data files:
   x <- read.nafo(year = 1960)
   
   # Read the 1960-1965 NAFO landings data files:
   x <- read.nafo(year = 1960:1965)

}
