% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.netmind.R
\name{read.netmind}
\alias{read.netmind}
\title{Read Netmind Acoustic Trawl Data.}
\usage{
read.netmind(x, file, offset = 0, repeats = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{year}{Survey year(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}
}
\description{
Functions to read Netmind acoustic trawl measurement data.
}
\examples{
# Read snow crab survey Netmind data from 1990 to 1994:
x <- read.netmind(1990:1994)

}
