% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.notus.R
\name{read.notus}
\alias{read.notus}
\title{Read Trawl Monitoring Data}
\usage{
read.notus(file)
}
\arguments{
\item{file}{Data file name.}
}
\description{
Read trawl acoustic monitoring data files. Acoustic sensors are attached to bottom trawls to measure
various aspects of the trawl, such as the door spread, wing spread, water depth and temperature.
}
\seealso{
read.scanmar read.netmind read.esonar
}
