% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nss.R
\name{read.nssset}
\alias{read.nssset}
\alias{read.nss}
\alias{read.nsscat}
\alias{read.nssbio}
\alias{read.nsslen}
\title{Read Northumberland Strait Survey Data}
\usage{
read.nssset(x, file, survey, ...)

read.nsscat(x, file, survey, species, ...)

read.nssbio(x, file, survey, species, ...)

read.nsslen(x, file, survey, species, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{species}{Species code or name.}

\item{year}{Survey year(s).}
}
\description{
Functions to Northumberland Strait survey  data sets.
}
\section{Functions}{
\itemize{
\item \code{read.nssset}: Read southern Gulf of Saint Lawrence Northumberland Strait survey set data.

\item \code{read.nsscat}: Read southern Gulf of Saint Lawrence Northumberland Strait survey catch data.

\item \code{read.nssbio}: Read southern Gulf of Saint Lawrence Northumberland Strait survey biological data.

\item \code{read.nsslen}: Read southern Gulf of Saint Lawrence Northumberland Strait survey length data.
}}

\examples{
 
# Read Northumberland Strait survey set data files:
x <- read.nssset()                 # Read all available data.
x <- read.nssset(year = 2019)      # Read single year.
x <- read.nssset(year = 2010:2015) # Read range of years.

# Read Northumberland Strait survey biological data files:
x <- read.nssbio(2020)  
x <- read.scsset(2020, species = "lobster")

}
\seealso{
\code{\link[gulf.data]{nss}}

\code{\link[gulf.data]{read.scs}}
}
