% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.oracle.R
\name{read.oracle}
\alias{read.oracle}
\title{Read Oracle Database}
\usage{
read.oracle(query, dsn, uid, password, believeNRows = FALSE, ...)
}
\arguments{
\item{query}{SQL query.}

\item{dsn}{Registered data source.}

\item{uid}{User identification.}

\item{password}{User password.}
}
\description{
Functions to access Oracle data.
}
