% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.probe.R
\name{read.minilog}
\alias{read.minilog}
\alias{read.star.oddi}
\alias{read.esonar}
\alias{read.notus}
\title{Read Probe Data.}
\usage{
read.minilog(x, file, offset = 0, repeats = FALSE, ...)

read.star.oddi(x, file, offset = 0, repeats = FALSE, ...)

read.esonar(x, file, offset = -3 * 60, repeats = FALSE, ...)

read.notus(file)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{year}{Survey year(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}
}
\description{
Functions to read probe data, such as depth/temperature or acoustic trawl monitoring data.
}
\section{Functions}{
\itemize{
\item \code{read.minilog}: Read Minilog probe data.

\item \code{read.star.oddi}: Read Star Oddi probe data.

\item \code{read.esonar}: Read \strong{eSonar} trawl acoustic monitoring data.

\item \code{read.notus}: \strong{Notus} trawl acoustic monitoring data.
}}

