% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.scsset}
\alias{read.scsset}
\alias{read.scsbio}
\alias{read.star.oddi}
\alias{read.esonar}
\alias{read.notus}
\alias{read.alsi}
\title{Read \strong{gulf} data sets.}
\usage{
read.scsset(x, file, ...)

read.scsbio(x, file, ...)

read.star.oddi(x, file, offset = 0, repeats = FALSE, ...)

read.esonar(x, file, offset = -3 * 60, repeats = FALSE, ...)

read.notus(file)

read.alsi(year, site, species, type, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{year}{Survey year(s).}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}
}
\description{
Functions to read Gulf of Saint-Lawrence data sets.
}
\section{Functions}{
\itemize{
\item \code{read.scsset}: Read southern Gulf of Saint Lawrence snow crab survey set data.

\item \code{read.scsbio}: Read southern Gulf of Saint Lawrence snow crab survey biological data.

\item \code{read.star.oddi}: Read Star Oddi probe data.

\item \code{read.esonar}: Read \strong{eSonar} trawl acoustic monitoring data.

\item \code{read.notus}: \strong{Notus} trawl acoustic monitoring data.

\item \code{read.alsi}: Read data from the Altantic Lobster Stock-recruitment Index. Larvae collectors are used to mainly provide recruitment indices for lobster and rock crab.
}}

\examples{
 
x <- read.alsi(year = 2010, species = 2550)
}
