% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.scobs.R
\name{read.scobs}
\alias{read.scobs}
\title{Read Observer Data}
\usage{
read.scobs(x, year, type = "sea", source = "R", ...)
}
\arguments{
\item{year}{Year.}

\item{type}{Sampling type ('sea' or 'port').}

\item{source}{Data source ('r', 'ascii', 'csv', or 'oracle').}

\item{file}{File name(s).}

\item{path}{Data file path.}

\item{cfvn}{Canadian Fishing Vessel Number(s).}

\item{trip.number}{Trip number identifier.}

\item{database}{Oracle database name.}

\item{username}{Oracle user name.}

\item{password}{Oracle password.}
}
\description{
Function to access and read at-sea and port observer data.
}
