% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.scs.R
\name{read.scsset}
\alias{read.scsset}
\alias{read.scs}
\alias{read.scscat}
\alias{read.scsbio}
\alias{read.scslen}
\title{Read Snow Crab Survey Data}
\usage{
read.scsset(x, file, survey, ...)

read.scscat(x, file, survey, species, ...)

read.scsbio(x, file, survey, category, drop = TRUE, verbose = FALSE, ...)

read.scslen(x, file, survey, species, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{species}{Species code or name.}

\item{drop}{Logical value specifying whether to remove empty data rows.}

\item{year}{Survey year(s).}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}
}
\description{
Functions to read snow crab survey data.
}
\section{Functions}{
\itemize{
\item \code{read.scsset}: Read southern Gulf of Saint Lawrence snow crab survey set data.

\item \code{read.scscat}: Read southern Gulf of Saint Lawrence snow crab survey by-catch data.

\item \code{read.scsbio}: Read southern Gulf of Saint Lawrence snow crab survey biological data.

\item \code{read.scslen}: Read southern Gulf of Saint Lawrence snow crab survey by-catch length data.
}}

\examples{
 
# Read snow crab survey set data files:
x <- read.scsset()                 # Read all available data.
x <- read.scsset(year = 2019)      # Read single year.
x <- read.scsset(year = 2010:2015) # Read range of years.

# Read specific tow data:
x <- read.scsset(2020, valid = 1)  # Load only valid tows.
x <- read.scsset(2020, tow.id = "GP354F")
x <- read.scsset(2020, date = "2020-07-13")
x <- read.scsset(2020, zone = "F")

# Using snow crab set data to specify corresponding biological data:
b <- read.scsbio(read.scsset(2020, valid = 1, zone = "F"))
b <- read.scsbio(2020, category = "MI")

}
\seealso{
\code{\link[gulf.data]{scs}}

\code{\link[gulf.data]{locate.scs}}

\code{\link[gulf.data]{read.nss}}
}
