% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.scuba.R
\name{read.scuba}
\alias{read.scuba}
\title{Read Scuba Survey Data}
\usage{
read.scuba(year, table, compress = TRUE, source = "dmapps")
}
\arguments{
\item{year}{Study year(s).}

\item{table}{Character string specifying which data table to read. Options are \sQuote{outings}, \sQuote{dives}, \sQuote{sections}, \sQuote{biological} or \sQuote{observations}.}

\item{compress}{Logical value specifying whether to automatically remove empty data columns or rows.}

\item{source}{Data source.}
}
\description{
Functions to access data from the SCUBA transect survey in the southern Gulf of Saint Lawrence.
}
\examples{
x <- read.scuba(2012:2014)  # Read all SCUBA data tables for the 2012 to 2014 seasons.
x <- read.scuba(2021)       # Read all SCUBA data tables for the 2021 season.

x <- read.scuba(2021, table = "section") # Read transect sectio data from the 2021 season.
}
