% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.seabird.R
\name{read.seabird}
\alias{read.seabird}
\alias{read.seabird.default}
\title{Read Seabird Probe Data.}
\usage{
read.seabird(x, ...)

\method{read.seabird}{default}(x, file, offset = 0, repeats = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{file}{File name(s).}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{year}{Survey year(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}
}
\description{
Functions to read Seabird data, such as depth/temperature or acoustic trawl monitoring data.
}
\section{Methods (by class)}{
\itemize{
\item \code{read.seabird(default)}: Read a Seabird data file.

}}
