% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.vms.R
\name{read.vms}
\alias{read.vms}
\title{Read Vessel Moniroting System (VMS) Data}
\usage{
read.vms(
  year,
  month = 1:12,
  dsn = options("gulf.oracle")[[1]]$vms$dsn,
  uid = options("gulf.oracle")[[1]]$vms$uid,
  password
)
}
\arguments{
\item{year}{Fishery year.}

\item{month}{Fishery month.}

\item{dsn}{Registered data source.}

\item{uid}{User identification.}

\item{password}{User password.}
}
\description{
Functions to access and update vessel monitoring system data.
}
