% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.winch.R
\name{read.winch}
\alias{read.winch}
\alias{read.winch.default}
\title{Read Trawl Winch Probe Data}
\usage{
read.winch(x, ...)

\method{read.winch}{default}(file)
}
\arguments{
\item{file}{File name.}
}
\description{
Read trawl winch speed and warp length data.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Read a trawl winch data file.
}}

