% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{reference.points}
\alias{reference.points}
\title{Harvest Control Rules}
\usage{
reference.points(species)
}
\arguments{
\item{species}{Species name or code.}
}
\description{
Calculate harvest control rates.
}
\examples{
x <- seq(0, 125000, len = 1000)
y <- TAC(x, species = "snow crab")
plot(range(x) / 1000, c(0, 0.5), type = "n", xlab = "Commercial biomass (x1000 t)", ylab = "Exploitation rate", yaxs = "i")
grid()
lines(x / 1000, y, lwd = 2)
box()
}
