% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Administrative regions}
\usage{
region(x, ...)
}
\arguments{
\item{x}{Character string containing region descriptors.}

\item{...}{Not used.}

\item{language}{Language to be returned.}

\item{output}{Output format, either \sQuote{region}, \sQuote{description} or \sQuote{abbreviation}.}
}
\description{
Functions to provide administrative regions identifiers and descriptions.
}
\examples{
# Complete sex code and description table:
region()

# Description for numeric sex codes:
region("g")

}
