% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\alias{sampler.default}
\alias{sampler.scs}
\alias{sampler.sco}
\title{Field Samplers}
\usage{
sampler(x, ...)

\method{sampler}{default}(x, project, ...)

\method{sampler}{scs}(x)

\method{sampler}{sco}(x)
}
\arguments{
\item{x}{Object containing field sampler information.}

\item{...}{Unused.}

\item{project}{Research project (see \code{\link{project}}).}
}
\description{
These are functions to clean up or standardize the names of field samplers from 
             various projects, such as research projects or sea-sampling.
}
\section{Functions and Methods}{

\describe{
    \item{\code{sampler}}{Generic \code{sampler} method.}
    \item{\code{sampler.default}}{Default sampler method.}
    \item{\code{sampler.scs}}{Field sampler clean-up function for snow crab survey.}
    \item{\code{sampler.sco}}{Field sampler clean-up function for snow crab at-sea and port 
                              observer data.}
}
}

