% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanmar.R
\name{scanmar}
\alias{scanmar}
\alias{scanmar.default}
\title{\strong{Scanmar} Data}
\usage{
scanmar(x, ...)

\method{scanmar}{default}(x, ...)
}
\arguments{
\item{x}{An \code{scanmar} object, data file, survey year or keyword search term.}

\item{header}{\strong{Scanmar} file header information to be assigned as metadata.}
}
\description{
Create \strong{scanmar} trawl acoustic monitoring probe data objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{scanmar} object.
}}

\examples{
locate.scanmar()

# Scanmar files for the 1990 snow crab survey:
locate.scanmar(year = 1990)

# Locate files with a specific tow ID from snow crab survey 2018-2020:
locate.scanmar("GP001", year = 2018:2020)

# Working example:
x <- read.scanmar("GP001", year = 1990)
describe(x)  # Description of file contents.
header(x)    # File header information.
plot(x)      # Graphical summary.
summary(x)   # Data summary.
}
