% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scobs.R
\name{scobs}
\alias{scobs}
\alias{scobs.default}
\title{Snow Crab Observer Data Class:}
\usage{
scobs(x, ...)

\method{scobs}{default}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object containing data fields to be mapped onto a corresponding snow crab observer data object.}

\item{...}{Other parameters (not used).}
}
\description{
Class containers for snow crab observer data.
}
\section{Methods (by class)}{
\itemize{
\item \code{scobs(default)}: Create an \code{scobs} object.

}}
