% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scs.R
\name{scsset}
\alias{scsset}
\alias{scscat}
\alias{scsbio}
\alias{scslen}
\alias{scsset.default}
\alias{scscat.default}
\alias{scsbio.default}
\alias{scslen.default}
\title{Snow Crab Survey Data Classes}
\usage{
scsset(x, ...)

scscat(x, ...)

scsbio(x, ...)

scslen(x, ...)

\method{scsset}{default}(x, ...)

\method{scscat}{default}(x, ...)

\method{scsbio}{default}(x, ...)

\method{scslen}{default}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object containing data fields to be mapped onto a corresponding snow crab survey data object.}

\item{...}{Other parameters (not used).}
}
\description{
Class containers for snow crab survey data.
}
\section{Functions}{
\itemize{
\item \code{scsset()}: Generic \code{scsset} method.

\item \code{scscat()}: Generic \code{scscat} method.

\item \code{scsbio()}: Generic \code{scsbio} method.

\item \code{scslen()}: Generic \code{scslen} method.

\item \code{scsset(default)}: Create an \code{scsset} object.

\item \code{scscat(default)}: Create an \code{scscat} object.

\item \code{scsbio(default)}: Create an \code{scsbio} object.

\item \code{scslen(default)}: Create an \code{scslen} object.

}}
\examples{
# Create 'scscat' object from a data frame:
x <- scscat(data.frame(date = "2020-07-19", tow.id = "GP001F", tow.number = 1, species = 10, number.caught = 144))
key(x)

# Create 'scsbio' object with specified 'tow.number' and 'sex' fields:
x <- scsbio(data.frame(date = "2020-07-19",
                       tow.id = "GP001F",
                       crab.number = 1:10,
                       carapace.width = rnorm(10, 100, 5), 
                       chela.height = NA,
                       sex = 1))

}
\seealso{
\code{\link{read.scs}}
}
