% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsbio.R
\name{scsbio}
\alias{scsbio}
\alias{scsbio.default}
\alias{scsbio.scsset}
\alias{read.scsbio}
\alias{update.scsbio}
\title{Snow Crab Biological Data}
\usage{
scsbio(x, ...)

\method{scsbio}{default}(x, format = fmt.scsbio(), ...)

\method{scsbio}{scsset}(x, ...)

read.scsbio(x, year, ...)

\method{update}{scsbio}(year, path, Rfile = TRUE, csv = TRUE, ...)
}
\arguments{
\item{x}{A \sQuote{data.frame} object. When reading data, \code{x} may be a numeric 
vector specifying the survey years to be loaded or an \code{\link{scsset}} object for which we want to
read the corresponding biological data.}

\item{...}{Other parameters (not used).}

\item{format}{An \sQuote{fmt} object which specifies the ASCII printing format for an 
\code{scbio} object.}

\item{year}{Numeric vector specifying the survey years to be loaded.}

\item{path}{string specifying the path to read or write snow crab biological data.}

\item{Rfile}{Logical value specifying whether to write an \code{R} data format file when 
updating snow crab biological data.}

\item{csv}{Logical value specifying whether to write a comma-separated (\code{csv}) format file 
when updating snow crab biological data.}

\item{key}{Character string(s) specifying which data columns form the index key.
This key can then be used as a default for such operations as merging data.}

\item{package}{Logical value specifying whether to write a comma-separated (\code{csv}) format
to the \code{gulf.data} package clone.}
}
\description{
The \code{scsbio} class is a container for snow crab biological data, i.e. 
             information about individual organisms sampled on the snow crab annual 
             survey.
}
\section{Functions}{

\describe{
  \item{\code{scsbio}}{Generic \code{scsbio} method.}
  \item{\code{scsbio.default}}{Create an \code{scsbio} object.}
  \item{\code{scsbio.scsset}}{Access snow crab biological data associated with snow crab survey set data.}
  \item{\code{read.scsbio}}{Read snow crab survey biological data.}
  \item{\code{update.scsbio}}{Update snow snow crab survey biological data repositories.}
}
}

\examples{
# Create empty 'scbio' object:
x <- scsbio()
   
# Create 'scbio' object with specified 'tow.number' and 'sex' fields:
x <- scsbio(data.frame(tow.number = 1:10, sex = 1))
   
x <- read.scsbio(year = 2012)   


}
