% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scscat.R
\name{scscat}
\alias{scscat}
\alias{scscat.default}
\title{Snow Crab Survey By-Catch Data}
\usage{
scscat(x, ...)

\method{scscat}{default}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object. When reading data, \code{x} may be a numeric vector specifying the 
survey years to be loaded or an \code{\link{scsset}} object for which we want to read the 
corresponding biological data.}

\item{...}{Other parameters (not used).}
}
\description{
The \code{scscat} class is a container for Snow Crab Survey by-catch data, i.e. information 
             about species other than snow crab caught on snow crab annual surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{scscat} object.
}}

\examples{
# Create 'scscat' object from a data frame:
x <- scscat(data.frame(date = "2020-07-19", tow.id = "GP001F", tow.number = 1, species = 10, number.caught = 144))
key(x)

# Read data:    
x <- read.scscat()                 # Read all available data.
x <- read.scscat(year = 2019)      # Read single year.
x <- read.scscat(year = 2010:2015, species = 10) # Load cod data for range of years.
x <- read.scscat(species = "Cod") # Load cod data using species name search for all years.

# Attach catches to tow data:
x <- read.scsset(2020, valid = 1, survey = "regular") # Read tow data.
y <- read.scscat(2020, species = "plaice")            # Read American plaice data.
import(x, var = "number.caught", fill = 0) <- y       # Attach number of specimens caught.

}
\seealso{
\code{\link{read.scsset}},  \code{\link{read.scscat}}
}
