% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabird.R
\name{seabird}
\alias{seabird}
\alias{seabird.default}
\title{\strong{Seabird} Data}
\usage{
seabird(x, ...)

\method{seabird}{default}(x, ...)
}
\arguments{
\item{x}{An \code{seabird} object, data file, survey year or keyword search term.}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{full.names}{Logical value specifying whether to include file paths when searching for \strong{seabird} data files.}

\item{header}{\strong{seabird} file header information to be assigned as metadata.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{remove}{Character string specifying keyword(s) which, if found in the data file or path, are removed from the search results.}
}
\description{
Functions to read, manipulate and analyze \strong{seabaird} probe data.
}
\section{Methods (by class)}{
\itemize{
\item \code{seabird(default)}: Default \code{seabird} method. Create an \code{seabird} object.

}}
