% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{sex}
\alias{sex}
\title{Sex Codes and Descriptions.}
\usage{
sex(x, ...)
}
\arguments{
\item{x}{Numerical or character vector containing sex code or descriptors.}

\item{language}{Language to be returned.}
}
\description{
Functions to provide sex code options and descriptions.

This function matches a sex code to a description of the sex of an animal.
If no \code{code} argument is given, then a named character vector of all
valid codes is returned.
}
\examples{
# Complete sex code and description table:
sex()

# Description for numeric sex codes:
sex(1)
sex(1, language = "fr")
sex(1:10) # Vector input.

# Mixture of character input formats:
sex(c("1", "2", "3", "m", "F", "mal", "fe", NA, "f", "femelle", "h", "female"), language = "french")
}
