% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.condition.R
\name{shell.condition}
\alias{shell.condition}
\title{Shell Condition}
\usage{
shell.condition(x, ...)
}
\arguments{
\item{x}{Numerical value(s) specifying shell condition code(s).}
}
\description{
Returns a description for snow crab shell codes. Unknown codes and NA values are returned as NA values.
}
\examples{
# Get the table of all defined shell condition codes:
shell.condition()
shell.condition(1)
shell.condition(0:2)
}
