% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.group.R
\name{is.fish}
\alias{is.fish}
\alias{is.invertebrate}
\alias{is.skate}
\alias{is.shrimp}
\title{Species Group Membership}
\usage{
is.fish(species)

is.invertebrate(species)

is.skate(species)

is.shrimp(species)
}
\arguments{
\item{species}{Numerical species code(s).}
}
\description{
Functions to check membership in various species groups.
}
\section{Functions}{

\describe{
  \item{\code{is.fish}}{Determine whether a species is an invertebrate.}
  \item{\code{is.invertebrate}}{Determine whether a species is an invertebrate.}
  \item{\code{is.skate}}{Determine whether a species is a skate.}
  \item{\code{is.shrimp}}{Determine whether a species was catgeroized as a shrimp historically.} 
}
}

