% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species}
\alias{species}
\alias{species.str}
\alias{species.default}
\alias{species.list}
\alias{species.numeric}
\alias{species.character}
\title{Species Names or Code}
\usage{
species(x, ...)

\method{species}{default}(x, ...)

\method{species}{list}(x, ...)

\method{species}{numeric}(x, language = "english", coding = "rv", source, ...)

\method{species}{character}(x, language = "english", input = "stacac", output = "rv", ...)
}
\arguments{
\item{x}{Numerical species code(s) or species name(s).}

\item{language}{Character string specifying the language in which species names are to be returned. Either 
\code{\sQuote{english}}, \code{\sQuote{french}} or \code{\sQuote{latin}}. Note that not all 
language and coding combinations are available.}

\item{coding, input, output}{Character string specifying the type of species coding being used. Options are the standard 
research survey coding \code{\sQuote{rv}} (or equivalently \code{\sQuote{research}}), the Statistical
Coordinating Committee for the Atlantic Coast (STACAC) coding \code{\sQuote{stacac}}, the North 
American Fisheries Organization (NAFO) coding \code{\sQuote{nafo}}, and the observer and commercial 
file name prefix coding \code{\sQuote{observer}} and \code{\sQuote{commercial}}.}

\item{oracle}{Logical value specifying whether to use the Oracle database as a source. Default is \code{FALSE}.}
}
\value{
Vector of character strings containing species names.
}
\description{
Converts species codes to corresponding species names and vice versa.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: \sQuote{species} default function.

\item \code{list}: Apply \code{species} function to list elements.

\item \code{numeric}: Convert species code(s) to corresponding species name(s).

\item \code{character}: Convert species name(s) to corresponding numeric code(s).
}}

\examples{
species(10)  # Code for "Atlantic cod".
species(c(10, 12, 40:43), language = "latin") # Latin names for common species.
species(101, coding = "nafo") 
species(145, coding = "stacac") 
species(41, coding = "com") # Commercial file prefix for Witch flounder:

# Search for two key words:
species("cod atl")

# Search for two key words:
v <- species(c("cod atl", "witch", "wint fl"))
species(v) # Extract names.

# Species with the word "Atlantic" in their name:
species(species("atlantic"))

}
\seealso{
\code{\link{species.table}}
}
