% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{speed}
\alias{speed}
\title{Survey Vessel Speed}
\usage{
speed(x, ...)
}
\arguments{
\item{x}{Data object containing trawl wing spread measurements.}

\item{...}{Not used.}
}
\description{
Functions to extract survey vessel speed.
}
\examples{
x <- read.scsset(2024, valid = 1)
speed(x)
}
