% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.number.R
\name{station.number.gulf.set}
\alias{station.number.gulf.set}
\title{Assign a Station Number}
\usage{
\method{station.number}{gulf.set}(
  x,
  method = "observed",
  distance.tolerance = 2,
  time.tolerance = 10,
  ...
)
}
\arguments{
\item{x}{A \sQuote{gulf.set} object.}

\item{method}{A character value specifying the method to used for fetching
or assigning station numbers. Accepted methods are either observed or
hard-coded values (\sQuote{observed}) or latitude-longitude coordinates
(\sQuote{calculated}, the default value).}

\item{distance.tolerance}{A positive numeric value stating the maximum
distance in kilometers beyond which proximate tows would not be considered
to be at a common sampling station. The default is \code{2}.}

\item{time.tolerance}{A positive numeric value stating the maximum time
interval in minutes beyond which proximate tows would not be considered to
be considered as comparative tows. The default is \code{10}.}

\item{list()}{Further arguments which are passed onto the
\code{\link[gulf.data]{station.number}} method.}
}
\value{
A numeric vector of station numbers for each set in \sQuote{x}.
}
\description{
Assigns a station number for each set in the September Research Vessel
survey.  This function groups sets into unique sampling sites. This function
is used to treat various types of repeated samples (e.g. repeat sets,
comparative sets, etc...)
}
\examples{

   # Read September comparative survey data for 2005:
   x <- read.gulf.set(year = 2021, password="R_GulfPKG_19")
   x <- x[which(x$experiment != 3),]

   # group the sampling stations using hard-coded fixed station assignment:
   x$set.number.observed <- station.number(x, method = "observed")

   # group the sampling stations using spatio-temporal proximity
   x$set.number.calculated <- station.number(x, method = "calculated", distance.tolerance = 2, time.tolerance=10)
   
}
