% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.number.R
\name{station.number}
\alias{station.number}
\alias{station.number.default}
\title{Sampling Station Numbering}
\usage{
station.number(x, ...)

\method{station.number}{default}(
  longitude = NULL,
  latitude = NULL,
  time = NULL,
  distance.tolerance = NULL,
  ...
)
}
\arguments{
\item{longitude, latitude}{Numeric vector in decimal degrees.}

\item{time}{Numeric vector or time (\sQuote{POSIX}) objects.}

\item{distance.tolerance}{Distance (in kilometers) below which a pair of points is considered to belong to the same sampling station.}

\item{time.tolerance}{Numeric value specifying the time-difference below which a pair of points is considered 
to have occured at the same time.}
}
\description{
Functions to extract or determine trawl sampling station identification numbers. 
             A sampling station is to be interpreted as a particular location, e.g. if two tows share
             a station number they can be considered as having been sampled in the same location. 
             Station numbers can be assigned or determined based on spatial proximity.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default 'station.number' method.
}}

\section{Methods}{

\describe{
   \item{\code{station.number}}{Generic \code{station.number} method.}
   \item{\code{station.number.default}}{Default \code{station.number} method.}
}
}

