% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{station}
\alias{station}
\alias{station.gulf.set}
\title{Sampling Station Numbering}
\usage{
station(x, ...)

\method{station}{gulf.set}(
  x,
  method = "observed",
  distance.tolerance = 2,
  time.tolerance = 10,
  ...
)
}
\arguments{
\item{x}{Data object.}

\item{...}{Other arguments (not used).}

\item{method}{Character string specifying the method to determine the station number. Options are \sQuote{observed} and \sQuote{calculated}.}

\item{distance.tolerance}{Distance (in kilometers) below which a pair of points is considered to belong to the same sampling station.}

\item{time.tolerance}{Numeric value specifying the time-difference below which a pair of points is considered to have occured at the same time.}
}
\description{
Functions to extract or determine trawl sampling station identification numbers.
             A sampling station is to be interpreted as a particular location, e.g. if two tows share
             a station number they can be considered as having been sampled in the same location.
             Station numbers can be assigned or determined based on spatial proximity.
}
\section{Methods (by class)}{
\itemize{
\item \code{station(gulf.set)}: Return survey sampling station identifier.

}}
