% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.logbook.R
\name{summary.logbook}
\alias{summary.logbook}
\title{Summarize Fishery Logbook Data}
\usage{
\method{summary}{logbook}(x, by = c("cfvn"), resolution)
}
\description{
Generates a summary of fishery logbook data.
}
