% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.scsset}
\alias{summary.scsset}
\alias{summary.scsbio}
\alias{summary.esonar}
\title{\code{gulf.data} Summaries}
\usage{
\method{summary}{scsset}(x, ...)

\method{summary}{scsbio}(x, category, ...)

\method{summary}{esonar}(x, year, trim = TRUE, ...)
}
\arguments{
\item{category}{Biological category string(s). See \code{\link{category}} for more details.}

\item{by}{Character string(s) specifying which variables to group by when summarizing.}

\item{weight}{Logical value specifying whether to return a summary by weights rather than counts.}
}
\description{
Functions that generate data summaries for various types of data objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Generate summaries for snow crab survey set data.

\item \code{scsbio}: Generate summaries for snow crab survey biological data.

\item \code{esonar}: Generate data summary for \strong{eSonar} data.
}}

\examples{
b <- read.scsbio(2020)
summary(b) # General summary.
summary(b, category = c("m", "f"), by = "tow.id") # Male and female catch summary.

}
