% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tac.R
\name{TAC}
\alias{TAC}
\alias{reference.points}
\alias{harvest.control.rule}
\alias{total.allowable.catch}
\title{Reference Points and Harvest Control Rules}
\usage{
reference.points(species)

harvest.control.rule(x, limit.reference.point, upper.stock.reference, species)

total.allowable.catch(x, ...)

TAC(x, ...)
}
\arguments{
\item{species}{Species name or code.}

\item{x}{Stock index value.}

\item{limit.reference.point}{Limit reference point value.}

\item{upper.stock.reference}{Upper stock reference point value.}
}
\description{
Calculate harvest control rates.
}
\examples{
x <- seq(0, 125000, len = 1000)
y <- TAC(x, species = "snow crab")
plot(range(x) / 1000, c(0, 0.5), type = "n", xlab = "Commercial biomass (x1000 t)", ylab = "Exploitation rate", yaxs = "i")
grid()
lines(x / 1000, y, lwd = 2)
box()
}
