% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{taxon}
\alias{taxon}
\title{Retrieve Taxonomic Information}
\usage{
taxon(x, ...)
}
\arguments{
\item{x}{Species code.}

\item{aphia.id}{Aphia taxonomic identification codes.}

\item{rank}{Taxonomic rank (e.g. "kingdom", "phylum", "genus", etc ...).}
}
\description{
Retrieve taxonomic information associated with identification codes.
}
\examples{
taxon(10:100)             # Latin genus + species names.
taxon(10:100, "phylum")   # Phylum names.
taxon(10:100, "family")   # Family names.
taxon(10:100, c("class", "order", "family")) # Class, order and family names.

}

