% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temperature}
\alias{temperature}
\title{Water Temperatures}
\usage{
temperature(x, ...)
}
\arguments{
\item{x}{Year.}

\item{year}{Year.}

\item{month}{Month of the year. Default is \sQuote{september}.}

\item{longitude, latitude}{Coordinates at which water temperatures are to be estimated. Units in decimal degrees.}

\item{depth}{Water depth at which temperatures are to be retrieved. Units are in meters. Alternatively, 
setting \code{depth = 'surface'} will return surface temperatures and setting \code{depth = 'bottom'}
will return bottom temperatures.}

\item{radius.search}{Search radius in meters used to average out temperatures within a range of the specied depth.}
}
\description{
Functions to extract of estimate water temperatures.
}
\examples{
# Load all september temperature data as a 3-D matrix:
z <- temperature(2019)

z <- temperature(2019, depth = "surface")

# Load snow crab survey data:
s <- read.scsset(2019, valid = 1, survey = "regular")

# Get june bottom temperatures using snow crab survey coordinates:
z <- temperature(2019, month = "June", depth = "bottom", longitude = lon(s), latitude = lat(s))

# Get september bottom temperatures using snow crab survey coordinates:
z <- temperature(2019, depth = "bottom", longitude = lon(s), latitude = lat(s))

# Read temperatures for the top 50 meters at the snow crab survey's coordinates:
z <- temperature(2019, depth = 0:50, longitude = lon(s), latitude = lat(s))
image(1:nrow(s), as.numeric(dimnames(z)$depth), z, xlab = "Sampling station", ylab = "Depth(m)")
}
