% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{start.time}
\alias{start.time}
\alias{stop.time}
\alias{touchdown.time}
\alias{liftoff.time}
\alias{haul.time}
\title{Trawl Event Times}
\usage{
start.time(x, ...)

stop.time(x, ...)

touchdown.time(x, ...)

liftoff.time(x, ...)

haul.time(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
These are functions for extracting the times at which various events occur during trawling. 
             For example, there are the start and stop times determined by on-board observers or trawl touchdown and 
             liftoff times.
}
\section{Functions}{
\itemize{
\item \code{start.time}: Generic \code{start.time} method. This is the time at which trawling is considered to have begun.

\item \code{stop.time}: Generic \code{stop.time} method. This is the time at which trawling is considered to have stopped.

\item \code{touchdown.time}: Generic \code{touchdown.time} method. This is the time at which the trawl makes its initial bottom contact.

\item \code{liftoff.time}: Generic \code{liftoff.time} method. This is the time at which the trawl makes is lifted off the bottom for hauling.

\item \code{haul.time}: Generic \code{haul.time} method. This is the time at which the trawl is considered to have been hauled.
}}

\examples{
x <- read.scsset(2020)
start.time(x)

x <- read.esonar(2020, tow.id = "GP354F")
start.time(x)
stop.time(x)
}
