% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time.scsset}
\alias{time.scsset}
\title{Trawl Event Times}
\usage{
\method{time}{scsset}(x, event = "start", ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
These are functions for extracting the times at which various events occur during trawling. 
             For example, there are the start and stop times determined by on-board observers or trawl touchdown and 
             liftoff times.
}
\examples{
x <- read.scsset(2020)
time(x, "start")

x <- read.esonar(2020, tow.id = "GP354F")
time(x, "start")
time(x, "stop")
}
