% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tow.id.R
\name{tow.id}
\alias{tow.id}
\alias{tow.id.default}
\alias{tow.id.probe}
\title{Extract Tow Identifier}
\usage{
tow.id(x, ...)

\method{tow.id}{default}(x, ...)

\method{tow.id}{probe}(x, method, max.distance = 500, ...)
}
\arguments{
\item{x}{Data object.}

\item{method}{Character string specifying the method to be used when determining the tow identification number for a 
\code{probe} object. Available methods are \sQuote{time} and \sQuote{latlong}.}

\item{max.distance}{Numeric valiue specifying the maximum distance tolerance (in kilometers) when determining the tow 
identification number for a \code{probe} object from lat-lon coordinates.}
}
\description{
Functions to extract or determine trawl tow identifiers.
}
\section{Methods}{

\describe{
   \item{\code{tow.id}}{Generic \code{tow.id} method.}
   \item{\code{tow.id.default}}{Default \code{tow.id} method.}
   \item{\code{tow.id.probe}}{Determine tow ID for a \code{probe} object.}
}
}

