% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tow.number.R
\name{tow.number}
\alias{tow.number}
\alias{tow.number.default}
\alias{tow.number.data.frame}
\title{Extract Tow Number}
\usage{
tow.number(x, ...)

\method{tow.number}{default}(x, ...)

\method{tow.number}{data.frame}(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{...}{Not used.}

\item{method}{Character string specifying the method to be used when determining the tow identification number for a 
\code{probe} object. Available methods are \sQuote{time} and \sQuote{latlong}.}

\item{max.distance}{Numeric valiue specifying the maximum distance tolerance (in kilometers) when determining the tow 
identification number for a \code{probe} object from lat-lon coordinates.}
}
\description{
Functions to extract or determine trawl tow numbers.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default 'tow.number' method.

\item \code{data.frame}: Default 'tow.number' method.
}}

\section{Methods}{

\describe{
   \item{\code{tow.number}}{Generic \code{tow.number} method.}
   \item{\code{tow.number.default}}{Default \code{tow.number} method.}
   \item{\code{tow.number.probe}}{Determine tow ID for a \code{probe} object.}
}
}

