% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\alias{trim.data.frame}
\alias{trim.probe}
\title{Trim Time Series Data}
\usage{
trim(x, ...)

\method{trim}{data.frame}(x, range, ...)

\method{trim}{probe}(x, start.time, end.time, buffer = 0, ...)
}
\arguments{
\item{x}{Data object containing time series data.}

\item{range}{Two-element vector specifying the time range of data beyond which data frame records will be removed. 
\code{NA} or \code{+/-Inf} can be used to specify open bounds.}

\item{start.time, end.time}{Time bounds beyind which probe data are to be truncated. If left unspecified, 
start and end times are determined from the project study data, such as set/tow data.}
}
\description{
Removes time series data which lie beyond a specified time interval.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Removes time series data which lie beyond a specified time interval for a \code{data.frame} object.

\item \code{probe}: Removes time series data which lie beyond start and end times for a \code{probe} object.
}}

