% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip.R
\name{trip}
\alias{trip}
\alias{trip.scsset}
\title{Trip Identifiers}
\usage{
trip(x)

\method{trip}{scsset}(x)
}
\arguments{
\item{x}{Object containing trip information.}
}
\description{
Return unique identifiers for each fishing trip in a data set.
}
\section{Methods}{

\describe{
   \item{\code{trip}}{Generic \code{trip} method.}
   \item{\code{trip.scsset}}{Determine snow crab survey trip using an analysis of date sequences.}
}
}

\examples{
x <- read.scsset(2019)
trip(x)

}
\seealso{
\code{\link{scsset}}
}
