% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\alias{units.default}
\alias{units<-}
\alias{units<-.default}
\title{Measurement Units}
\usage{
units(x, ...)

\method{units}{default}(x, ...)

units(x, ...) <- value

\method{units}{default}(x, y, ...) <- value
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned measurement units.}

\item{value}{Character string(s) specifying the units to be assigned.}
}
\value{
An object with a \code{units} attribute attached to it.
}
\description{
Assigns or retrieves measurement units assigned to observed data.
}
\details{
The \code{units} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{

\describe{
  \item{\code{units}}{Generic \code{units} extraction method.}
  \item{\code{units.default}}{Default \code{units} extraction method.}
  \item{\code{units<-}}{Generic \code{units} assignment method. See Examples for usage.}
  \item{\code{units<-.default}}{Default \code{units} assignment method. See Examples for usage.}
}
}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
units(x, "measurement") <- "centimeters"

# Show attributes:
attributes(x)

# Erase 'units' attribute:
units(x) <- NULL

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{format}}
}
