% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.ziff.R
\name{update.ziff}
\alias{update.ziff}
\title{Update ZIFF Data}
\usage{
\method{update}{ziff}(year, region = "g", ...)
}
\arguments{
\item{year}{Year.}

\item{region}{Fishing zone identifier.}

\item{...}{Not used.}
}
\description{
Update Zonal Interchange File Format (ZIFF) data, by reading ASCII data and saving an R data version.
}
