% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.R
\name{vessel}
\alias{vessel}
\alias{vessel.default}
\alias{vessel.character}
\title{Project Identifiers}
\usage{
vessel(x, ...)

\method{vessel}{default}(x, ...)

\method{vessel}{character}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Character search string.}

\item{...}{Other arguments (not used).}
}
\value{
An object with a \code{project} attribute attached to it.
}
\description{
Functions to retrieve survey vessel information.
}
\section{Methods}{

\describe{
   \item{\code{vessel}}{Generic \code{vessel} method.}
   \item{\code{vessel.default}}{Default \code{vessel} method. Returns the \code{vessel.csv} data table.}
   \item{\code{vessel.character}}{Search for vessel name and return vessel specifications.}
}
}

\examples{
vessel()               # Survey vessel table.
vessel("prince")       # Search for vessel name.
}
