% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\alias{weight.default}
\alias{weight.scsbio}
\alias{weight.scobs}
\title{Individual or Catch Weight}
\usage{
weight(x, ...)

\method{weight}{default}(x, species, sex, coefficients, units = "kg", ...)

\method{weight}{scsbio}(x, category, by, as.hard.shelled, units = "g", ...)

\method{weight}{scobs}(x, ...)
}
\arguments{
\item{x}{A numerical vector of organism lengths, or alternatively, a frequency table or a 
data frame of length-frequencies as produced by \code{\link{freq}}. The presence 
of \sQuote{species}, \sQuote{sex} or \sQuote{year} fields in the data frame will 
be passed onto the corresponding function arguments.}

\item{...}{Futher arguments.}

\item{species}{Species code.}

\item{sex}{Numerical code specifying sex.}

\item{coefficients}{A two-element numerical vector specifying the \code{a} and \code{b} 
coefficients, respectively. The the \sQuote{a} coefficient is assumed 
to be on the log-10 scale and the units in grams.}

\item{units}{Units of the weight vector to be returned. It may be either in grams 
(\code{units = "g"}) or kilograms (\code{units = "kg"}).}

\item{category}{A character string specifying a snow crab or crustacean category for syntax.}

\item{by}{Character string(s) specifying the fields by which to group the estimated weights.}

\item{year}{Survey years to use as data when calculating the length-weight coefficients. 
If left unspecified, a table of default values are used.}

\item{probability}{Logical value specifying whether maturity values are to be filled in with 
probabilities when morphometric values are unavailable. In this case, the 
numbers returned may be non-integer values.}
}
\value{
Returns a numerical vector the same size as \code{length} containing the expected weight 
        of an organism for a given length measurement.
}
\description{
Returns an estimate of the weight of individual organisms for a specified length. 
             Estimated weights of sample subsets can also be calculated.
             
             This function makes use of a set of allometric length-weight coefficients to
             estimate the weight of an organism given its length. The species must be
             specified and the sex may be specified for dimorphic species.

             If \code{year} is specified, then survey data is loaded and the length-weight 
             coefficients are calculated directly.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default weight function.

\item \code{scsbio}: Weight function for \code{scsbio} objects.

\item \code{scobs}: Weight function for \code{scobs} objects.
}}

\examples{
# Weights for Atlantic cod:
weight(0:100, species = 10)

# Weights for female white hake:
weight(0:100, species = 12, sex = 2)

# Weights for female white hake based on 2010 September survey data:
weight(0:100, species = 12, sex = 2, year = 2010)

# Weights for white hake based on pooled sexes and data from 2010-2013 September surveys:
weight(0:100, species = 12, sex = 2, year = 2010:2013)

# Transform length-frequencies to weight-length:
x <- read.gulf(year = 2014, species = 40, card = "len")

# Simple example:
f <- freq(x, scale = TRUE)  # Pooled length-frequencies.
weight(f, species = 40)

# Length-frequencies by stratum and sex:
f <- freq(x, scale = TRUE, by = c("species", "stratum", "sex"))
weight(f)

# Length-frequencies by stratum and sex, use RV 2014 length-eight coefficients:
f <- freq(x, scale = TRUE, by = c("species", "stratum", "sex", "year"))
weight(f)

# Load 2010 snow crab data:
x <- read.scsbio(year = 2012)

# Calculate weight for each crab:
weight(x)

# Calculate weights by tow:
weight(x, by = "tow.number")

# Calculate total weights by day:
weight(x, by = c("year", "month", "day"), category = c("TM", "TMM", "TMSC12", "TMSC345"))
}
