% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingspread.R
\name{wingspread}
\alias{wingspread}
\alias{wingspread.esonar}
\alias{wingspread.scsset}
\title{Trawl Wing Spread}
\usage{
wingspread(x, ...)

\method{wingspread}{esonar}(x, ...)

\method{wingspread}{scsset}(x, ...)
}
\arguments{
\item{x}{Data object containing trawl wing spread measurements.}

\item{...}{Not used.}
}
\description{
Functions to extract trawl wing spread measurements.
}
\section{Methods (by class)}{
\itemize{
\item \code{wingspread(esonar)}: Fetch trawl wing spread measures from an \code{esonar} object.

\item \code{wingspread(scsset)}: Calculate the average wingspread for a snow crab survey tow.

}}
\examples{
x <- read.esonar(2020)
wingspread(x)

x <- read.scsset(2024, valid = 1)
wingspread(x)
}
