% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\title{Attach Catch Data}
\usage{
catch(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{value}{Object containing catch data to be assigned.}
}
\description{
Attach catch data.
}
\section{Functions}{

\describe{
   \item{\code{catch}}{Generic \code{catch} method.}
   \item{\code{catch<-}}{Generic \code{catch} assignment method.}
   \item{\code{catch<-.scsset}}{Import catch data into an \code{\link{scsset}} object.}
}
}

