% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch<-}
\alias{catch<-}
\alias{catch<-.scsset}
\title{Attach Catch Data}
\usage{
catch(x, ...) <- value

\method{catch}{scsset}(x, ...) <- value
}
\arguments{
\item{x}{Target object.}
}
\description{
Attach catch data.
}
\section{Functions}{

\describe{
  \item{\code{catch}}{Generic \code{catch} assignment method.}
  \item{\code{catch<-.scsset}}{Import catch data into an \code{\link{scsset}} object. }
}
}

