% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esonar.R
\name{esonar}
\alias{esonar}
\title{\strong{eSonar} Data}
\usage{
esonar(x, ...)
}
\arguments{
\item{x}{An \code{esonar} object, data file, survey year or keyword search term.}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{full.names}{Logical value specifying whether to include file paths when searching for \strong{eSonar} data files.}

\item{header}{\strong{eSonar} file header information to be assigned as metadata.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep repeated data records.}

\item{remove}{Character string specifying keyword(s) which, if found in the data file or path, are removed from the search results.}
}
\description{
Functions to read, manipulate and analyze \strong{eSonar} trawl acoustic monitoring  probe data.
}
\section{Functions}{

\describe{
   \item{\code{esonar}}{Generic \code{esonar} method.}
   \item{\code{esonar.default}}{Create an \code{esonar} object.}
   \item{\code{locate.esonar}}{Find \code{esonar} data file(s).}
   \item{\code{read.esonar}}{Read \code{esonar} data file(s).}
   \item{\code{plot.esonar}}{Plot \code{esonar} data.}
   \item{\code{map.esonar}}{Map \code{esonar} vessel and data track.}
   \item{\code{describe.esonar}}{\code{esonar} data description.}
   \item{\code{summary.esonar}}{Summary statistics for an \code{esonar} object.}
   \item{\code{match.esonar}}{Match \code{esonar} data records.}
   \item{\code{truncate.esonar}}{Remove \code{esonar} which lie outside a specified time interval.}
}
}

\examples{
# eSonar files for the 2020 snow crab survey:
locate.esonar(year = 2020)

# Locate files with a specific tow ID from snow crab survey 2018-2020:
locate.esonar("GP001", year = 2018:2020)

# Working example:
x <- read.esonar("GP001", year = 2020)
describe(x)  # Description of file contents.
header(x)    # File header information.
plot(x)      # Graphical summary.
summary(x)   # Data summary.

}
