% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esonar.R
\name{esonar}
\alias{esonar}
\alias{esonar.default}
\alias{locate.esonar}
\alias{read.esonar}
\alias{plot.esonar}
\alias{map.esonar}
\alias{describe.esonar}
\alias{summary.esonar}
\alias{match.esonar}
\alias{truncate.esonar}
\title{\strong{eSonar} Data}
\usage{
esonar(x, ...)

\method{esonar}{default}(x, header, ...)

\method{locate}{esonar}(x, year, tow.id, full.names = TRUE, remove = "test", ...)

read.esonar(x, offset = 0, repeats = FALSE, ...)

\method{plot}{esonar}(x, ...)

map.esonar(x, set.card = NULL, variable = NULL, ...)

\method{describe}{esonar}(x, ...)

\method{summary}{esonar}(x, year, truncate = TRUE, round = TRUE, ...)

match.esonar(x, set.card, method = "file.name")

\method{truncate}{esonar}(x, buffer = 0, ...)
}
\arguments{
\item{x}{An \code{esonar} object, data file, survey year or keyword search term.}

\item{header}{\strong{eSonar} file header information to be assigned as metadata.}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{full.names}{Logical value specifying whether to include file paths when searching for \strong{eSonar} data files.}

\item{remove}{Character string specifying keyword(s) which, if found in the data file or path, are removed from the search results.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep repeated data records.}
}
\description{
Functions to read, manipulate and analyze \strong{eSonar} trawl acoustic monitoring  probe data.
}
\section{Functions}{

\describe{
   \item{\code{esonar}}{Generic \code{esonar} method.}
   \item{\code{esonar.default}}{Create an \code{esonar} object.}
   \item{\code{locate.esonar}}{Find \code{esonar} data file(s).}
   \item{\code{read.esonar}}{Read \code{esonar} data file(s).}
   \item{\code{plot.esonar}}{Plot \code{esonar} data.}
   \item{\code{map.esonar}}{Map \code{esonar} vessel and data track.}
   \item{\code{describe.esonar}}{\code{esonar} data description.}
   \item{\code{summary.esonar}}{Summary statistics for an \code{esonar} object.}
   \item{\code{match.esonar}}{Match \code{esonar} data records.}
   \item{\code{truncate.esonar}}{Remove \code{esonar} which lie outside a specified time interval.}
}
}

\examples{
# eSonar files for the 2020 snow crab survey:
locate.esonar(year = 2020)

# Locate files with a specific tow ID from snow crab survey 2018-2020:
locate.esonar("GP001", year = 2018:2020)

# Working example:
x <- read.esonar("GP001", year = 2020)
describe(x)  # Description of file contents.
header(x)    # File header information.
plot(x)      # Graphical summary.
summary(x)   # Data summary.

}
