% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{grow}
\alias{grow}
\title{Project Growth}
\usage{
grow(x, ...)
}
\arguments{
\item{x}{Specimen size or length-frequency.}

\item{x0}{}

\item{n}{Numeric value(s) specifying the number of iterations to apply the growth functions. In crustaceans, 
the number of iterations can be thought as the number of moults. Non-integer values are allowed.}
}
\value{
If \code{x} is left unspecified, then a function is returned which can be used to 
evaluate growth if given inputs.
}
\description{

}
