% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.data.R
\name{gulf.data}
\alias{gulf.data}
\alias{.onLoad}
\title{Gulf Data Package}
\usage{
.onLoad(libname, pkgname)
}
\description{
\strong{gulf.data} is an \code{R} package for accessing southern Gulf of Saint Lawrence fisheries science data.
}
