% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key.scsset}
\alias{key.scsset}
\alias{key.nssset}
\alias{key.nsscat}
\alias{key.nssbio}
\alias{key.nsslen}
\alias{key.scscat}
\alias{key.scsbio}
\alias{key.scslen}
\alias{key.star.oddi}
\alias{key.esonar}
\title{\strong{gulf.data} Index Keys}
\usage{
\method{key}{scsset}(x, ...)

\method{key}{nssset}(x, ...)

\method{key}{nsscat}(x, ...)

\method{key}{nssbio}(x, ...)

\method{key}{nsslen}(x, ...)

\method{key}{scscat}(x, ...)

\method{key}{scsbio}(x, ...)

\method{key}{scslen}(x, ...)

\method{key}{star.oddi}(x, ...)

\method{key}{esonar}(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Define or extract index keys for various \code{gulf.data} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{key(scsset)}: Fetch \code{scsset} index key.

\item \code{key(nssset)}: Index key for Northumberland Strait catch set (\code{nsscat}).

\item \code{key(nsscat)}: Index key for Northumberland Strait catch data (\code{nsscat}).

\item \code{key(nssbio)}: Index key for Northumberland Strait biological data (\code{nssbio}).

\item \code{key(nsslen)}: Index key for Northumberland Strait length data (\code{nsslen}).

\item \code{key(scscat)}: Index key for snow crab survey by-catch data (\code{scscat}).

\item \code{key(scsbio)}: Fetch \code{scsbio} index key.

\item \code{key(scslen)}: Fetch \code{scslen} index key.

\item \code{key(star.oddi)}: Fetch \code{star.oddi} index key.

\item \code{key(esonar)}: Fetch \code{esonar} index key.

}}
\examples{
x <- read.scsset(2020)
key(x)
key.scsset()
}
