% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key.scsset}
\alias{key.scsset}
\alias{key.nssset}
\alias{key.nsscat}
\alias{key.scscat}
\alias{key.scsbio}
\alias{key.star.oddi}
\alias{key.esonar}
\title{\strong{gulf.data} Index Keys}
\usage{
\method{key}{scsset}(x, ...)

\method{key}{nssset}(x, ...)

\method{key}{nsscat}(x, ...)

\method{key}{scscat}(x, ...)

\method{key}{scsbio}(x, ...)

\method{key}{star.oddi}(x, ...)

\method{key}{esonar}(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Define or extract index keys for various \code{gulf.data} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Fetch \code{scsset} index key.

\item \code{nssset}: Index key for Northumberland Strait catch set (\code{nsscat}).

\item \code{nsscat}: Index key for Northumberland Strait catch data (\code{nsscat}).

\item \code{scscat}: Index key for snow crab survey by-catch data (\code{scscat}).

\item \code{scsbio}: Fetch \code{scsbio} index key.

\item \code{star.oddi}: Fetch \code{star.oddi} index key.

\item \code{esonar}: Fetch \code{esonar} index key.
}}

\examples{
x <- read.scsset(2020)
key(x)
key.scsset()
}
