% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.probe.R
\name{locate.probe}
\alias{locate.probe}
\alias{locate.minilog.default}
\alias{locate.minilog.scsset}
\alias{locate.star.oddi.default}
\alias{locate.star.oddi.scsset}
\alias{locate.scanmar.default}
\alias{locate.netmind.default}
\alias{locate.esonar.default}
\alias{locate.esonar.scsset}
\title{Locate Probe Data}
\usage{
\method{locate}{probe}(
  x,
  probe,
  project = "scs",
  location = "headline",
  remove,
  year,
  tow.id,
  ...
)

\method{locate}{minilog.default}(x, ...)

\method{locate}{minilog.scsset}(x, ...)

\method{locate}{star.oddi.default}(x, ...)

\method{locate}{star.oddi.scsset}(x, ...)

\method{locate}{scanmar.default}(x, ...)

\method{locate}{netmind.default}(x, ...)

\method{locate}{esonar.default}(x, ...)

\method{locate}{esonar.scsset}(x, ...)
}
\arguments{
\item{x}{Study year, data file name or tow identification tag.}

\item{probe}{Probe name. See \code{\link[gulf.data]{probe}} for available options.}

\item{project}{Project name. See \code{\link[gulf.data]{project}} for available options.}

\item{location}{Probe location. For trawl probes, options are \code{headline} and \code{footrope}.}

\item{remove}{Character string specifying terms to remove from search results.}

\item{year}{Study year.}

\item{tow.id}{Tow identification tag for trawl studies and surveys.}
}
\description{
Locate data files for various types of probes.
}
\section{Functions}{
\itemize{
\item \code{locate.minilog.default}: Default method for locating Minilog probe data files.

\item \code{locate.minilog.scsset}: Locate Minilog associated with snow crab survey tow data.

\item \code{locate.star.oddi.default}: Default method for locating Star Oddi probe data files.

\item \code{locate.star.oddi.scsset}: \code{scsset} method for locating Star Oddi data files.

\item \code{locate.scanmar.default}: Default method for locating Scanmar acoustic data files.

\item \code{locate.netmind.default}: Default method for locating Netmind acoustic data files.

\item \code{locate.esonar.default}: Default method for locating eSonar data files.

\item \code{locate.esonar.scsset}: \code{scsset} method for locating eSonar data files.
}}

\examples{
locate.probe(2020, "esonar", project = "scs")                             # 2020 Snow crab survey eSonar data files.
locate.probe(2018, "minilog", project = "scs")                            # 2018 snow crab survey Minilog data files.
locate.probe(2018, "star.oddi", project = "scs", location = "headline")   # 2018 snow crab survey Star Oddi data files.
locate.probe(probe = "scanmar", project = "scs")                          # All snow crab survey Scanmar files.

locate.minilog()                     # Find all available Minilog data files.
locate.minilog(1997)                 # Find Minilog data files for the 1997 snow crab survey.
locate.minilog(1997:1999)            # Find Minilog data files for the 1997-1997 snow crab survey.

locate.star.oddi(2020, probe = "headline") 
locate.star.oddi(2020, probe = "tilt", tow.id = "GP354F")

#' # Global searches:
locate.scanmar()     # Find all available Scanmar data files.
locate.scanmar(1990) # Find Scanmar data files for the 1990 snow crab survey.
locate.scanmar(1990:1994) # Find Scanmar data files for the 1990-1994 snow crab survey.

# Specific searches:
locate.scanmar(1990, tow.id = 100)
locate.scanmar(1990, tow.id = "100")
locate.scanmar(1990, tow.id = "S90100")
locate.scanmar(tow.id = "S90100")

locate.netmind()     # Find all available Netmind data files.
locate.netmind(1999) # Find Netmind data files for the 1990 snow crab survey.
locate.netmind(1999:2004) # Find Netmind data files for the 1990-1994 snow crab survey.

# Specific searches:
locate.netmind(1999, tow.id = 100)
locate.netmind(1999, tow.id = "100")
locate.netmind(1999, tow.id = "S90100")
locate.netmind(tow.id = "355")
 
locate.esonar(2020) # Find snow crab survey eSonar data from 2020.
locate.esonar(2020, tow.id = "GP354F")

}
\seealso{
\code{\link[gulf.metadata]{project}}, \code{\link[gulf.data]{probe}}
}
