% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate.scsset}
\alias{locate.scsset}
\alias{locate.scsbio}
\title{Locate Gulf Data Files}
\usage{
\method{locate}{scsset}(x, year, source = "gulf.data", ...)

\method{locate}{scsbio}(x, year, source = "gulf.data", remove = "bad", ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Study or survey year.}

\item{source}{Character string specifying the data source for data files. Options are \sQuote{ascii} or a gulf 
package name (e.g. \sQuote{gulf.data}).}

\item{remove}{Character string vector specifying a list of key words, which if found in the data path or file name, 
are removed from the search results.}

\item{tow.id}{Character string(s) specifying a tow identifier(s) (e.g. \sQuote{GP354F}).}

\item{probe}{Character string specifying probe type or location. For example \sQuote{headline} specifies the Star Oddi 
probe attached to the trawl headline on snow crab surveys.}
}
\description{
Functions to locate various data file(s) from different studies and surveys.
}
\section{Methods (by class)}{
\itemize{
\item \code{scsset}: Locate snow crab survey set data files.

\item \code{scsbio}: Locate snow crab survey biological data files.
}}

\examples{
locate.scsset()      # All snow crab survey set data files.
locate.scsset(2020)  # Load specific year(s).

locate.esonar(2020) # Find snow crab survey eSonar data from 2020.
locate.esonar(2020, tow.id = "GP354F")

locate.star.oddi(2020, probe = "headline", source = 'ascii') 
locate.star.oddi(2020, probe = "footrope", source = 'ascii', tow.id = "GP354F")

}
