% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minilog.R
\name{minilog}
\alias{minilog}
\alias{minilog.default}
\alias{minilog.data.frame}
\alias{path.minilog}
\alias{file.minilog}
\alias{read.minilog}
\alias{write.minilog}
\alias{plot.minilog}
\alias{summary.minilog}
\alias{split.minilog}
\alias{truncate.minilog}
\alias{info.minilog}
\title{Minilog Data}
\usage{
minilog(x, ...)

\method{minilog}{default}(x, ...)

\method{minilog}{data.frame}(x, header, ...)

path.minilog(year, project, survey, ...)

file.minilog(
  x,
  year,
  set.number,
  tow.id,
  set.card,
  path = TRUE,
  survey = "sc",
  project,
  ...
)

read.minilog(x, project, offset = 0, file, ...)

write.minilog(x, file = NULL)

\method{plot}{minilog}(x, col, lwd = 1, ...)

\method{summary}{minilog}(x, truncate = TRUE, round = TRUE, ...)

\method{split}{minilog}(x, by = "file.name", set.card, start, end, buffer = 0)

\method{truncate}{minilog}(x, buffer = 0, ...)

info.minilog(x, set.card = NULL)
}
\arguments{
\item{x}{Minilog data file}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{survey, project}{Character string specifying the survey or project name. 
The survey may refer to the September multi-species survey (\code{survey = "rv"},
\code{survey = "sep"} or \code{survey = "September"}), the  Northumberland Strait 
survey (\code{= "ns"}), the mobile Sentinel survey  \code{= "sentinel"}),
or the snow crab survey (\code{= "sc"} or \code{= "snow crab"}).}

\item{set.number, tow.id}{Numeric value or strings specifying the survey set number or tow ID.}

\item{set.card}{Data frame containing the survey tows which specify the Minilog data to be loaded.}

\item{path}{Logical value specifying whether to include the path in the Minilog files.}
}
\description{
Functions to create, read, write, manipulate and analyze Minilog depth/temperature 
             probe data.
}
\section{Functions}{

\describe{
   \item{\code{minilog}}{Generic \code{minilog} method.}
   \item{\code{minilog.default}}{Default minilog method}
   \item{\code{minilog.data.frame}}{Convert data frame to minilog object.}
   \item{\code{path.minilog}}{Find a minilog data path.}
   \item{\code{file.minilog}}{Find a minilog data file.}
   \item{\code{read.minilog}}{Read Minilog data. }
   \item{\code{write.minilog}}{Write a \code{minilog} object to a file.}
   \item{\code{plot.minilog}}{Plot minilog data object.}
   \item{\code{summary.minilog}}{Minilog data object summary.}
   \item{\code{split.minilog}}{Split minilog data object into disjoint subsets.}
   \item{\code{truncate.minilog}}{Truncate minilog data object.}
   \item{\code{info.minilog}}{Minilog data object information.}
}
}

\examples{
# Minilog files for the 2000 September RV survey:
minilog.file(survey = "rv", year = 2000)

# Use a set card extract for the first ten sets of 2012:
x <- read.gulf(year = 2012, survey = "sc")
minilog.file(x[1:10, ])

# Use a tow ID to extract file names for the snow crab survey 2006-2012:
minilog.file("GP001", year = 2006:2012)

x <- read.minilog(tow.id = "GP001F", year = 2010)
x <- read.minilog(survey = "rv", year = 2010)

# Plot Minilog data:
file <- system.file("extdata", "Minilog example.txt", package = "gulf.data")
x <- read.minilog(file)
plot(x)

}
