% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minilog.R
\name{minilog}
\alias{minilog}
\alias{minilog.default}
\alias{minilog.data.frame}
\title{Minilog Data}
\usage{
minilog(x, ...)

\method{minilog}{default}(x, ...)

\method{minilog}{data.frame}(x, header, ...)
}
\arguments{
\item{x}{Minilog data file}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{survey, project}{Character string specifying the survey or project name. 
The survey may refer to the September multi-species survey (\code{survey = "rv"},
\code{survey = "sep"} or \code{survey = "September"}), the  Northumberland Strait 
survey (\code{= "ns"}), the mobile Sentinel survey (\code{= "sen"} or \code{= "sentinel"}),
or the snow crab survey (\code{= "sc"} or \code{= "snow crab"}).}

\item{set.number, tow.id}{Numeric value or strings specifying the survey set number or tow ID.}

\item{set.card}{Data frame containing the survey tows which specify the Minilog data to be loaded.}

\item{path}{Logical value specifying whether to include the path in the Minilog files.}
}
\description{
Minilog probes are used to record depth and temperature data. 
These are functions to create, read and write \code{minilog} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{minilog} object.

\item \code{data.frame}: Converts a data frame to a \code{minilog} object.
}}

\examples{
# Minilog files for the 2000 September RV survey:
minilog.file(survey = "rv", year = 2000)

# Use a set card extract for the first ten sets of 2012:
x <- read.gulf(year = 2012, survey = "sc")
minilog.file(x[1:10, ])

# Use a tow ID to extract file names for the snow crab survey 2006-2012:
minilog.file("GP001", year = 2006:2012)

x <- read.minilog(tow.id = "GP001F", year = 2010)
x <- read.minilog(survey = "rv", year = 2010)

}
\seealso{
\code{\link[gulf.data]{read.minilog}}

\code{\link[gulf.data]{header}}

\code{\link[gulf.data]{plot.minilog}}
}
