% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minilog.R
\name{minilog}
\alias{minilog}
\alias{minilog.default}
\title{\strong{Minilog} Data}
\usage{
minilog(x, ...)

\method{minilog}{default}(x, ...)
}
\arguments{
\item{x}{An \code{minilog} object, data file, survey year or keyword search term.}

\item{year}{Numeric value specifying the survey or project year(s).}

\item{full.names}{Logical value specifying whether to include file paths when searching for \strong{minilog} data files.}

\item{header}{\strong{minilog} file header information to be assigned as metadata.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}

\item{remove}{Character string specifying keyword(s) which, if found in the data file or path, are removed from the search results.}
}
\description{
Functions to read, manipulate and analyze \strong{minilog} trawl acoustic monitoring  probe data.
}
\section{Methods (by class)}{
\itemize{
\item \code{minilog(default)}: Default \code{minilog} method. Create an \code{minilog} object.

}}
\examples{
# minilog files for the 2020 snow crab survey:
locate.minilog(year = 2020)

# Locate files with a specific tow ID from snow crab survey 2018-2020:
locate.minilog("GP001", year = 2018:2020)

# Working example:
x <- read.minilog(year = 2018, tow.id = "GP354F")
describe(x)  # Description of file contents.
header(x)    # File header information.
summary(x)   # Data summary.

}
\seealso{
\link{locate.probe}, \link{read.minilog}
}
