% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moulting.R
\name{is.new.shell}
\alias{is.new.shell}
\alias{is.new.shell.scsbio}
\alias{is.hard.shell}
\alias{is.hard.shell.scsbio}
\alias{is.hard.shell.scsobs}
\alias{is.soft.shell}
\alias{is.skip.moulter}
\alias{is.skip.moulter.scsbio}
\title{Carapace Moult Status}
\usage{
is.new.shell(x, ...)

\method{is.new.shell}{scsbio}(x)

is.hard.shell(x, ...)

\method{is.hard.shell}{scsbio}(x, durometer, shell.condition, zone, ...)

\method{is.hard.shell}{scsobs}(x, durometer, shell.condition, zone, ...)

is.soft.shell(x)

is.skip.moulter(x, ...)

\method{is.skip.moulter}{scsbio}(x, ...)
}
\arguments{
\item{x}{Object.}
}
\value{
Logical vector.
}
\description{
Determines the carapace moult or hardness status.
}
\section{Functions}{

\describe{
  \item{\code{is.new.shell}}{Returns whether a crustacean is newly moulted.}
  \item{\code{is.new.shell.scsbio}}{Returns whether a crab is newly moulted for snow crab survey data.}
  \item{\code{is.hard.shell}}{Returns whether a carapace is considered hard.}
  \item{\code{is.hard.shell.scsbio}}{Returns whether a carapace is considered hard for snow crab survey data.}
  \item{\code{is.hard.shell.scsobs}}{Returns whether a carapace is considered hard for snow crab observer data.}
  \item{\code{is.soft.shell}}{Returns whether a carapace is considered soft.}
  \item{\code{is.skip.moulter}}{Returns whether a crustacean is a skip-moulter.}
  \item{\code{is.skip.moulter.scsbio}}{Returns whether a snow crab is a skip-moulter.}
}
}

\examples{
# Read 2010 crab data:
x <- read.scsbio(2010)
index <- is.new.shell(x)    # Newly moulted crab.
index <- is.multiparous(x)  # Multiparous females.
index <- is.skip.moulter(x) # Skip-moulters.

}
