% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe.R
\name{probe}
\alias{probe}
\alias{probe.default}
\title{Probe Data Class}
\usage{
probe(x, ...)

\method{probe}{default}(x, header, ...)
}
\arguments{
\item{x}{Probe data object.}

\item{header}{File header data.}

\item{buffer}{Extra time, in seconds, to be included beyond the start and end time specifications when truncating data.}
}
\description{
Measurement probe data class definition and methods.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create a \code{probe} class object.
}}

\section{Methods}{

\describe{
   \item{\code{probe}}{Generic \code{probe} method.}
   \item{\code{probe.default}}{Create a \code{probe} object.}
}
}

\seealso{
\code{\link{tow.id}}
}
