% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.colorimeter.R
\name{read.colorimeter}
\alias{read.colorimeter}
\alias{locate.colorimeter}
\title{Locate and Read Colorimetry Data}
\usage{
read.colorimeter(...)

\method{locate}{colorimeter}(x, year, source = "gulf.data", ...)
}
\arguments{
\item{x}{Data object.}

\item{year}{Survey year.}

\item{tow.id}{Character string(s) specifying a tow identifier(s) (e.g. \sQuote{GP313F}).}
}
\description{
Functions to find and access colour meter data from the snow crab survey.
}
\section{Functions}{
\itemize{
\item \code{locate(colorimeter)}: Locate snow crab colorimetry files.

}}
\examples{
# Locate data files:
locate.colorimeter()          # Find all snow crab colorimeter data files.
locate.colorimeter(2020)      # Find specific year.
locate.colorimeter(2017:2019) # Find set of years.

# Read data:
x <- read.colorimeter()
x <- read.colorimeter(2020)
x <- read.colorimeter(tow.id = "GP313F")

}
\seealso{
\code{\link{read.colorimeter}}
}
