% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gulf.set.R
\name{read.gulf.set}
\alias{read.gulf.set}
\title{Read Survey Set Data}
\usage{
read.gulf.set(
  year,
  survey = "rv",
  dsn = options("gulf.oracle")[[1]]$rvs$dsn,
  uid = options("gulf.oracle")[[1]]$rvs$uid,
  ...
)
}
\arguments{
\item{year}{Survey year.}

\item{survey}{Survey name.}

\item{dsn}{Data source name.}

\item{uid}{User identification.}

\item{password}{Oracle database password.}
}
\description{
Read trawl set data from various groundfish, crustacean and pelagic surveys.
}
