% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.logbook.R
\name{read.logbook}
\alias{read.logbook}
\title{Load Fishery Logbook Data}
\usage{
read.logbook(x, species = "snow crab", ...)
}
\arguments{
\item{x}{Fishery year or file name.}

\item{year}{Fishery year.}

\item{file}{File name to be loaded. File format is assumed to be comma-separated (i.e. .csv).}

\item{path}{File path containing comma-separated fishery logbook files (i.e. .csv).}
}
\description{
Function to load logbook and slip landings data from the southern Gulf of Saint Lawrence fisheries.
}
