% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.notus.R
\name{read.notus}
\alias{read.notus}
\title{Notus Acoustic Trawl Data}
\usage{
read.notus(file)
}
\arguments{
\item{file}{Data file name.}
}
\description{
Read Notus trawl acoustic monitoring data files. Acoustic sensors are attached to bottom trawls 
             to measure various aspects of the trawl, such as the door spread, wing spread, water depth and 
             temperature.
}
\seealso{
\code{\link{probe}}, \code{\link{esonar}}, \code{\link{netmind}}, \code{\link{scanmar}}
}
