% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.notus.R
\name{read.notus}
\alias{read.notus}
\title{Read Notus Acoustic Trawl Data.}
\usage{
read.notus(x, file, echo = FALSE, ...)
}
\arguments{
\item{x}{Survey year or file name.}

\item{file}{File name(s).}

\item{...}{Other parameters passed onto \code{locate} functions or used to subset data.}

\item{year}{Survey year(s).}

\item{survey}{Survey type, as determined by the \link{survey.scsset} function.}

\item{tow.id}{Numeric value or character string specifying the ID corresponding to a particular tow sampling station.}

\item{offset}{Numeric value specifying the offset time (in minutes) to include as a corrective in the data time stamps.}

\item{repeats}{Logical value specifying whether to keep or average out data records with identical time stamps.}
}
\description{
Functions to read Notus acoustic trawl data, such as depth/temperature or acoustic trawl monitoring data.
}
