% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsbio.R
\name{scsbio}
\alias{scsbio}
\alias{scsbio.default}
\title{Snow Crab Survey Biological Data}
\usage{
scsbio(x, ...)

\method{scsbio}{default}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} object. When reading data, \code{x} may be a numeric 
vector specifying the survey years to be loaded or an \code{\link{scsset}} object for which we want to
read the corresponding biological data.}

\item{...}{Other parameters (not used).}
}
\description{
The \code{scsbio} class is a container for Snow Crab Survey Biological data, i.e. information 
             about individual organisms sampled on the snow crab annual survey.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{scsbio} object.
}}

\examples{
# Create 'scsbio' object with specified 'tow.number' and 'sex' fields:
x <- scsbio(data.frame(tow.number = 1:10, sex = 1))

# Read data:    
x <- read.scsbio()                 # Read all available data.
x <- read.scsbio(year = 2019)      # Read single year.
x <- read.scsbio(year = 2010:2015) # Read range of years.

summary(x)
}
