% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsset.R
\name{scsset}
\alias{scsset}
\alias{scsset.default}
\title{Snow Crab Set/Tow Data}
\usage{
scsset(x, ...)

\method{scsset}{default}(x, ...)
}
\arguments{
\item{x}{A \sQuote{data.frame} object. When reading data, \code{x} may be a numeric
vector specifying the survey years.}

\item{year}{Survey year(s) to be loaded.}
}
\description{
The \code{scsset} class is a container for Snow Crab Survey Set/Tow data, i.e.
             information about individual tows performed on the snow crab annual survey.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Create an \code{scsset} object.
}}

\examples{
# Locate snow crab survey set data files:
locate.scsset()
locate.scsset(1988)
locate.scsset(2010:2020)
# Read snow crab survey set data files:
x <- read.scsset()                 # Read all avaliable data.
x <- read.scsset(year = 2019)      # Read single year.
x <- read.scsset(year = 2010:2015) # Read range of years.

}
\seealso{
\code{\link{scsbio}}, \code{\link{tow.id}}
}
