% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{species}
\alias{species}
\alias{species.foreign}
\alias{species.default}
\alias{species.list}
\alias{species.numeric}
\alias{species.character}
\title{Species Names or Code}
\usage{
species(x, ...)

\method{species}{foreign}(x, ...)

\method{species}{default}(x, ...)

\method{species}{list}(x, ...)

\method{species}{numeric}(x, language = "english", coding = "code", output, ...)

\method{species}{character}(x, language = "english", coding = "code", drop = TRUE, ...)
}
\arguments{
\item{x}{Numerical species code(s) or species name(s).}

\item{language}{Character string specifying the language in which species names are to be returned. Either 
\code{\sQuote{english}}, \code{\sQuote{french}} or \code{\sQuote{latin}}. Note that not all 
language and coding combinations are available.}

\item{coding, input, output}{Character string specifying the type of species coding being used. Options are the standard 
research survey coding \code{\sQuote{rv}} (or equivalently \code{\sQuote{research}}), the Statistical
Coordinating Committee for the Atlantic Coast (STACAC) coding \code{\sQuote{stacac}}, the North 
American Fisheries Organization (NAFO) coding \code{\sQuote{nafo}}, World Repository of Marine Species (WoRMS) 
coding \code{\sQuote{WoRMS}} or \code{\sQuote{aphiaID}},  and the observer and commercial 
file name prefix coding \code{\sQuote{observer}} and \code{\sQuote{commercial}}.}
}
\value{
Vector of character strings containing species names. For multiple species name searches, a list of
        search results are returned.
}
\description{
Converts species codes to corresponding species names and vice versa.
}
\section{Functions}{

\describe{
  \item{\code{species}}{Generic \code{species} method.}
  \item{\code{species.foreign}}{Load foreign species code table.}
  \item{\code{species.default}}{Default \code{species} method.}
  \item{\code{species.list}}{Apply \code{species} function to list elements.}
  \item{\code{species.numeric}}{Convert numeric species code(s) to corresponding species name(s).}
  \item{\code{species.character}}{Find numeric species code(s).}
  \item{\code{species.str}}{Alias of \code{species.numeric}.}
  \item{\code{species.code}}{Alias of \code{species.character}.}
}
}

\examples{
# Miscellaneous queries:
species(10)  # Code for "Atlantic cod".
species(c(10, 12, 40:43), language = "latin") # Latin names for common species.
species(101, coding = "nafo")     # NAFO species codes.
species(145, coding = "stacac")   # STACAC species codes.
species(41, coding = "com")       # Commercial file prefix for Witch flounder.
species(126436, coding = "WoRMS") # WoRMS species codes.

# Converting between species codes:
species(10, output = "WoRMS") # DFO to WoRMS coding.
species(126436, coding = "worms", output = "code")
species(10, output = "stacac")
species(10, output = "stacac")
species(10, output = "nafo")
species(101, coding = "nafo", output = "stacac")

# Search for two key words:
species("cod atl")

# Search for two key words:
v <- species(c("cod atl", "witch", "wint fl"))
species(v) # Extract names.

# Species entries with the word "Atlantic" in their name:
species(species("atlantic"))

# Extract entire species tables:
species()
species.foreign()

}
\seealso{
\code{\link{data}}
}
