% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.number.R
\name{station.number}
\alias{station.number}
\alias{station.number.default}
\title{Return a Station Number}
\usage{
station.number(x, ...)

\method{station.number}{default}(
  longitude = NULL,
  latitude = NULL,
  time = NULL,
  distance.tolerance = NULL,
  ...
)
}
\arguments{
\item{longitude}{A numeric vector. A coordinate vector in decimal degrees.}

\item{latitude}{A numeric vector. A coordinate vector in decimal degrees.}

\item{time}{A numeric vector or a vector of a vector of \code{R} time
(\sQuote{POSIX}) objects.}

\item{distance.tolerance}{A numeric value specifying the distance (in
kilometers) below which a pair of points is considered to belong to the same
group or station.}

\item{time.tolerance}{A numeric value specifying the time-difference below
which a pair of points is considered to belong to the same group or station.
The units are in minutes if the \code{time} argument belongs to a
\sQuote{POSIX} class.}
}
\description{
Functions to extract or determine trawl tow numbers.
}
\details{
Partition a set of points into groups based on 
spatial and/or temporal proximity.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default 'station.number' method.
}}

\section{Methods}{

\describe{
   \item{\code{station.number}}{Generic \code{station.number} method.}
   \item{\code{station.number.default}}{Default \code{station.number} method.}
}
}

