% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{Survey Identifiers}
\usage{
survey(x, ...)
}
\arguments{
\item{year}{Survey year(s).}

\item{survey}{Character string specifying the research survey. The input is passed onto the
\code{\link{project}} function. To see the complete list of survey IDs available,
run \code{survey()}.}

\item{verbose}{Logical value specifying whether to return the survey identifiers in long form.}
}
\description{
Returns research survey codes and identifiers.
}
\section{Functions}{

\describe{
  \item{\code{survey}}{Generic \code{survey} method.}
  \item{\code{survey.default}}{Fetch survey identifiers.}
  \item{\code{survey.character}}{Fetch survey identifiers and information.}
}
}

\examples{
# Complete lists survey codes:
survey()               # List all survey IDs.
survey(verbose = TRUE) # List survey full names.
survey("ns")           # Northumberland Strait survey ID.
survey("northumb")     # Northumberland Strait survey ID.
survey("sen")          # Sentinel survey ID.

# Return the September research vessel survey code for 2005:
survey(year = 2005:2007, survey = "rv")

# Return the September research vessel survey codes for a vector of years:
survey(survey = "rv", year = 2000:2019)

# Return the Nothumberland Strait survey codes for a vector of years:
survey(year = 2000:2009, survey = "ns")

# Return the juvenile cod survey codes for a vector of years:
survey(year = 1990:1995, survey = "juv")

}
