% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\alias{survey.default}
\alias{survey.scsset}
\alias{survey.nssset}
\title{Survey Identifiers}
\usage{
survey(x, ...)

\method{survey}{default}(x, year, project, ...)

\method{survey}{scsset}(x, ...)

\method{survey}{nssset}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{year}{Survey year(s).}

\item{survey}{Character string specifying the research survey. The input is passed onto the
\code{\link{project}} function. To see the complete list of survey IDs available,
run \code{survey()}.}

\item{verbose}{Logical value specifying whether to return the survey identifiers in long form.}
}
\description{
Returns research survey codes and identifiers.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Fetch survey identifiers.

\item \code{scsset}: Determine type of survey sampling for snow crab survey tows.

\item \code{nssset}: Determine type of survey sampling for Northumberland Strait survey tows.
}}

\examples{
# Complete lists survey codes:
survey()                      # Complete survey table.
survey(project = "sept")      # September multispecies survey table.
survey(project = "ns")        # Northumberland Strait survey ID.
survey(project = "northumb")  # Northumberland Strait survey ID.

survey("^P")                  # Surveys whose IDs start with "P".

# Return the September research vessel survey code for 2005:
survey(year = 2005:2007, project = "rv")

# Return the September research vessel survey codes for a vector of years:
survey(survey = "rv", year = 2000:2019)

# Return the Nothumberland Strait survey codes for a vector of years:
survey(year = 2000:2009, survey = "ns")

# Return the juvenile cod survey codes for a vector of years:
survey(year = 1990:1995, survey = "juv")

# Snow crab survey type:
x <- read.scsset(2019)
survey(x) # Separates regular fall and comparative survey sets.
}
