% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.R
\name{vessel}
\alias{vessel}
\alias{vessel.default}
\alias{vessel.character}
\alias{vessel.scsset}
\alias{vessel.gulf.set}
\title{Project Identifiers}
\usage{
vessel(x, ...)

\method{vessel}{default}(x, ...)

\method{vessel}{character}(x, verbose = FALSE, ...)

\method{vessel}{scsset}(x, ...)

\method{vessel}{gulf.set}(x, ...)
}
\arguments{
\item{x}{Character search string.}

\item{...}{Other arguments (not used).}
}
\description{
Functions to retrieve survey vessel information.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{vessel} method. Returns the complete vessel data table.

\item \code{character}: Search for vessel name and return vessel specifications.

\item \code{scsset}: Vessel names for snow crab survey.

\item \code{gulf.set}: Vessel names for various science surveys.
}}

\examples{
vessel()               # Survey vessel table.
vessel("prince")       # Search for vessel name.
vessel("marco")        # Search for vessel name.
vessel("opilio")       # Search for vessel name.

# Read snow crab set data:
x <- read.scsset(2000:2021)
vessel(x)  # Determine vessel for each data record.
}
