% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel.R
\name{vessel}
\alias{vessel}
\alias{vessel.default}
\alias{vessel.character}
\title{Project Identifiers}
\usage{
vessel(x, ...)

\method{vessel}{default}(x, ...)

\method{vessel}{character}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Character search string.}

\item{...}{Other arguments (not used).}
}
\description{
Functions to retrieve survey vessel information.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{vessel} method. Returns the complete vessel data table.

\item \code{character}: Search for vessel name and return vessel specifications.
}}

\examples{
vessel()               # Survey vessel table.
vessel("prince")       # Search for vessel name.
vessel("marco")        # Search for vessel name.
vessel("opilio")       # Search for vessel name.
}
