% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wingspread.R
\name{wingspread}
\alias{wingspread}
\alias{wingspread.esonar}
\title{Trawl Wing Spread}
\usage{
wingspread(x, ...)

\method{wingspread}{esonar}(x, ...)
}
\arguments{
\item{x}{Data object containing trawl wing spread measurements.}
}
\description{
Functions to extract trawl wing spread measurements.
}
\section{Methods (by class)}{
\itemize{
\item \code{esonar}: Fetch trawl wing spread measures from an \code{esonar} object.
}}

\examples{
x <- read.esonar(2020)
wingspread(x)
}
